% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{album_sales}
\alias{album_sales}
\title{Album sales data}
\format{
A tibble with 200 rows and 5 variables.
}
\source{
\href{https://www.discovr.rocks/csv/album_sales.csv}{www.discovr.rocks/csv/album_sales.csv}
}
\usage{
album_sales
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictitious data that imagines a world where I have a cool job in the music industry. Except, it's not \emph{that} cool because
my job is to predict album sales (broadly defined in some way that accounts for physical sales, streams and digital sales). In my
little fantasy I collect data from 200 releasures (albums). For each one, I have information about the amount spent
advertising the album, the number of sales, the number of plays on radio songs from the album had per week, and a rating of the image of the band.
The (fictional) data contains the following variables:
\itemize{
\item \strong{album_id}: album identifier.
\item \strong{adverts}: advertising budget in thousands of whatever currency is used in your country.
\item \strong{sales}: the number of album sales (physical, digital, streams)
\item \strong{airplay}: the number of times songs from the album were played on radio thew eek before release
\item \strong{image}: a rating of the band's image from scale from 0 (dad dancing at a disco) to 10 (sicker than a dog that's eaten a bag of onions)
}
}
\keyword{datasets}
