% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{angry_pigs}
\alias{angry_pigs}
\title{Video games and aggression example 1}
\format{
A tibble with 336 rows and 4 variables
}
\source{
\href{https://www.discovr.rocks/csv/angry_pigs.csv}{www.discovr.rocks/csv/speed_date.csv}
}
\usage{
angry_pigs
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Angry Birds is a video game in which you fire birds at pigs.  A (fabricated) study was set up in which people
played Angry Birds and a control game (Tetris) over a 2-year period (1 year per game).
They were put in a pen of pigs for a day before the study, and after 1 month, 6 months and 12 months.
Their violent acts towards the pigs were counted. The (fictional) data contains
\itemize{
\item \strong{id}: participant ID
\item \strong{game}: whether the participant had been assigned to play angry pigs or tetris
\item \strong{time}: the time at which aggressive acts were measured (Baseline, 1 month, 6 months and 12 months)
\item \strong{aggression}: the number of aggressive acts towards pigs
}
}
\keyword{datasets}
