% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{animal_bride}
\alias{animal_bride}
\title{Goat or dog data}
\format{A tibble with 20 rows and 3 variables.}
\source{
\url{https://www.discoveringstatistics.com/books/discovering-statistics-using-r/}
}
\usage{
animal_bride
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictitious data inspired by two news stories that I enjoyed. The first was about a Sudanese man who was forced to marry a goat (\url{http://ow.ly/9DyyP})
after being caught having sex with it. I’m not sure he treated the goat to a nice dinner in a posh restaurant before taking advantage of her, but either way you have to feel sorry for the goat.
I’d barely had time to recover from that story when another appeared about an Indian man forced to marry a dog(\url{http://ow.ly/9DyFn}) to atone for stoning two dogs and stringing them up in a tree 15 years earlier.
Why anyone would think it’s a good idea to enter a dog into matrimony with a man with a history of violent behaviour towards dogs is beyond me.
Still, I wondered whether a goat or dog made a better spouse. I found (but not really) some other people who had been forced to marry goats and dogs and measured their life satisfaction and, also, how much they like animals.
The data contains the following variables:

\itemize{
\item \strong{wife}: whether the person married a goat or a dog
\item \strong{animal}: how much the person likes animals
\item \strong{life_satisfaction}: the person's life satisfaction score
\item \strong{wife}: Whether the person married a goat or a dog
\item \strong{animal}: How much the person likes animals
\item \strong{life_satisfaction}: The person's life satisfaction score
}
}
\keyword{datasets}
