% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{animal_dance}
\alias{animal_dance}
\title{Dancing cats and dogs data}
\format{
A tibble with 270 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/animal_dance.csv}{www.discovr.rocks/csv/animal_dance.csv}
}
\usage{
animal_dance
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about dancing cats and dogs. A researcher was interested in whether animals could be trained to dance.
He took 200 cats and 70 dogs and tried to train them to line-dance by giving them either food or affection as a reward for dance-like behaviour.
At the end of the week he counted how many animals could line-dance and how many could not.
The object contains the following variables:
\itemize{
\item \strong{id}: participant ID
\item \strong{training}: factor describing whether the animal was trained using food of affection as a reward
\item \strong{dance}: factor describing whether the animal danced or not
\item \strong{animal}: factor describing whether the animal was a cat or a dog
}
}
\keyword{datasets}
