% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{anovaTest}
\alias{anovaTest}
\title{Anova Test}
\usage{
anovaTest(input)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
analyzes the variance of the given samples
}
\description{
Anova Test
}

