% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{band_personality}
\alias{band_personality}
\title{Musicians and personality data}
\format{
A tibble with 434 rows and 6 variables.
}
\source{
\href{https://www.discovr.rocks/csv/band_personality.csv}{www.discovr.rocks/csv/band_personality.csv}
}
\usage{
band_personality
}
\description{
A fictional dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A musicologist was interested in whether the instrument played in a rock band could be predicted by their personality traits.
She measured personality traits in 434 musicians and noted the instrument they played. See also \link{sing_or_guitar}.
The data contains the following variables:
\itemize{
\item \strong{id}: Participant's ID
\item \strong{neurotic}: participant's neuroticism score (high score = high neuroticism)
\item \strong{extroversion}: participant's extroversion score (high score = high extroversion)
\item \strong{agreeableness}: participant's agreeableness score (high score = high agreeableness)
\item \strong{conscientiousness}: participant's conscientiousness score (high score = high conscientiousness)
\item \strong{instrument}: the instrument played (singer, guitar, bass or drums)
}
}
\keyword{datasets}
