% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{bernard_2012}
\alias{bernard_2012}
\title{Bernard, et al. (2012) data}
\format{
A tibble with 78 rows and 6 variables.
}
\source{
\href{https://www.discovr.rocks/csv/bernard_2012.csv}{www.discovr.rocks/csv/bernard_2012.csv}
}
\usage{
bernard_2012
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
There is a concern that images that portray women as sexually desirable objectify them. Philippe Bernard and colleagues tested this idea in an inventive study that used the 'inversion effect', which describes the phenomenon
that people find it harder to recognize upside-down (inverted) pictures than ones the right way up. It turns out that this 'inversion effect' occurs for pictures of humans but not objects, so by seeing how easily people recognize
inverted sexualized pictures of women, Bernard could see whether these pictures were processed more like people or objects (Bernard et al., 2012). Participants were presented with pictures of sexualized (i.e., not wearing many clothes)
males and females, half of which were inverted (\code{inverted_women} and \code{inverted_men}) whereas the remainder were upright (\code{upright_women} and \code{upright_men}). Participants self-reported their gender. After each trial,
participants were shown two pictures and asked to identify the one they had just seen. The outcome was the proportion of correctly identified pictures.

An inversion effect is demonstrated by higher recognition scores for upright pictures than for inverted ones. If sexualized females are processed as objects, you would expect an inversion effect for the male pictures but not the female ones.
The data contains the following variables:
\itemize{
\item \strong{id}: Participant ID number
\item \strong{gender}: Self-reported gender
\item \strong{inverted_women}: the proportion of correctly identified inverted pictures of women
\item \strong{upright_women}: the proportion of correctly identified upright pictures of women
\item \strong{inverted_men}: the proportion of correctly identified inverted pictures of men
\item \strong{upright_men}: the proportion of correctly identified upright pictures of men
}
}
\references{
\itemize{
\item Bernard, P., Gervais, S. J., Allen, J., Campomizzi, S., & Klein, O. (2012). Integrating sexual objectification with object versus person recognition: The sexualized-body-inversion hypothesis. \emph{Psychological Science}, 23(5), 469–471. \doi{10.1177/0956797611434748}
}
}
\keyword{datasets}
