% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{big_hairy_spider}
\alias{big_hairy_spider}
\title{Big hairy spider data}
\format{
A tibble with 24 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/big_hairy_spider.csv}{www.discovr.rocks/csv/big_hairy_spider.csv}
}
\usage{
big_hairy_spider
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Is arachnophobia (fear of spiders) specific to real spiders or will pictures of spiders evoke similar levels
of anxiety? Twelve arachnophobes were asked to play with a big hairy tarantula with big fangs and an evil
look in its eight eyes and at a different point in time were shown only photos of the same spider.
The participants' anxiety was measured in each case. The (fictional) data contains the following variables:
\itemize{
\item \strong{id}: the participant's first name
\item \strong{spider_type}: whether the spider stimulus was a real spider or a photo of a spider
\item \strong{anxiety}: the participant's anxiety after exposure to the stimulus
}
}
\keyword{datasets}
