% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{biggest_liar}
\alias{biggest_liar}
\title{The biggest liar data}
\format{
A tibble with 68 rows and 4 variables.
}
\source{
\href{https://www.discoveringstatistics.com/books/discovering-statistics-using-r/}{https://www.discoveringstatistics.com/books/discovering-statistics-using-r}
}
\usage{
biggest_liar
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data based on the World’s Biggest Liar competition held annually at the Santon Bridge Inn in Wasdale (in the Lake District, UK).
Each year locals are encouraged to attempt to tell the biggest lie in the world. I wanted to test a theory that more creative people will be able to create taller tales.
I gathered together 68 past contestants from this competition and noted where they were placed in the competition (first, second, third, etc.);
I also gave them a creativity questionnaire (maximum score 60). The data set has four variables
\itemize{
\item \strong{id}: Participant id
\item \strong{creativity}: Creativity score (maximum score 60)
\item \strong{position}: position in competition as a numeric variable from 1 (first place) to 5 (fifth place)
\item \strong{novice}: factor coding whether this was the participant's first time in the competition (\emph{first time}) or if they had entered before (\emph{previous entrant}).
}
}
\keyword{datasets}
