% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{bronstein_miss_2019}
\alias{bronstein_miss_2019}
\title{Bronstein et al. (2019) data with missing values inserted}
\format{
A tibble with 947 rows and 5 variables
}
\source{
\href{https://www.discovr.rocks/csv/bronstein_miss_2019.csv}{www.discovr.rocks/csv/bronstein_miss_2019.csv}
}
\usage{
bronstein_miss_2019
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A version of the Bronstein et al. (2019) fake news data (\link{bronstein_2019}) but with missing values inserted using MCAR amputation (with the help of the \code{mice} package and \code{ampute()} function).
For details of variables see \link{bronstein_2019}.
}
\references{
\itemize{
\item Bronstein, M. V., Pennycook, G., Bear, A., Rand, D. G., & Cannon, T. D. (2019). Belief in fake news is associated with delusionality, dogmatism, religious fundamentalism, and reduced analytic thinking. \emph{Journal of Applied Research in Memory and Cognition}, 8(1), 108–117. \url{https://doi.org/10.1016/j.jarmac.2018.09.005}
}
}
\keyword{datasets}
