% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{burnout}
\alias{burnout}
\title{Burnout data}
\format{
A tibble with 467 rows and 7 variables.
}
\source{
\href{https://www.discovr.rocks/csv/burnout.csv}{www.discovr.rocks/csv/burnout.csv}
}
\usage{
burnout
}
\description{
A fictional dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about burnout. Four hundred and sixty-seven lecturers completed questionnaire measures of \code{burnout} (burnt out or not),
perceived \code{control} (high score = low perceived control), \code{coping} ability (high score = high ability to cope with stress),
stress from \code{teaching} (high score = teaching creates a lot of stress for the person),
stress from \code{research} (high score = research creates a lot of stress for the person) and stress from providing \code{pastoral} care
(high score = providing pastoral care creates a lot of stress for the person).
The data contains the following variables:
\itemize{
\item \strong{id}: Participant's ID
\item \strong{burnout}: whether the participant was classified as burnt out or not
\item \strong{control}: perceived control (high score = low perceived control)
\item \strong{coping}: coping ability (high score = high ability to cope with stress)
\item \strong{teaching}: stress from teaching (high score = teaching creates a lot of stress for the person)
\item \strong{research}: stress from research (high score = research creates a lot of stress for the person)
\item \strong{pastoral}: stress from providing pastoral care (high score = pastoral care creates a lot of stress for the person)
}
}
\keyword{datasets}
