% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{cat_dance}
\alias{cat_dance}
\title{Dancing cats data}
\format{
A tibble with 200 rows and 2 variables.
}
\source{
\href{https://www.discovr.rocks/csv/cat_dance.csv}{www.discovr.rocks/csv/cat_dance.csv}
}
\usage{
cat_dance
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about dancing cats. A researcher was interested in whether animals could be trained to dance.
He took 200 cats and tried to train them to line-dance by giving them either food or affection as a reward for dance-like behaviour.
At the end of the week he counted how many animals could line-dance and how many could not.
The object contains the following variables:
\itemize{
\item \strong{reward}: factor describing whether the cat was trained using food of affection as a reward
\item \strong{dance}: factor describing whether the cat danced or not
}
}
\keyword{datasets}
