% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{cat_reg}
\alias{cat_reg}
\title{Cat regression data}
\format{
A tibble with 200 rows and 7 variables.
}
\source{
\href{https://www.discovr.rocks/csv/cat_reg.csv}{www.discovr.rocks/csv/cat_reg.csv}
}
\usage{
cat_reg
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data illustrating how the chi-square test is a linear model. It's about line dancing cats. The object contains the following variables:
\itemize{
\item \strong{reward}: whether the cat was trained using food (0) of affection (1) as a reward
\item \strong{dance}: Whether the cat danced (1) or not (0)
\item \strong{interaction}: the interaction of dance and reward (i.e. dance multiplied by reward)
\item \strong{observed}: the observed frequency for the combination of dance and reward
\item \strong{expected}: the expected frequency for the combination of dance and reward
\item \strong{lmobserved}: the natural logarithm of the observed frequency for the combination of dance and reward
\item \strong{lmexpected}: the natural logarithm of the expected frequency for the combination of dance and reward
}
}
\keyword{datasets}
