% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{catterplot}
\alias{catterplot}
\title{Catterplot data}
\format{
A tibble with 78 rows and 2 variables.
}
\source{
\href{https://www.discovr.rocks}{www.discovr.rocks}
}
\usage{
catterplot
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data for plotting a catterplot. The object contains the following variables:
\itemize{
\item \strong{dinner_time}: the time (hours) since the cat was last fed
\item \strong{meow}: How loud the cat's purr is
}
}
\keyword{datasets}
