% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{child_aggression}
\alias{child_aggression}
\title{Child aggression data}
\format{
A tibble with 666 rows and 6 variables.
}
\source{
\href{https://www.discovr.rocks/csv/child_aggression.csv}{www.discovr.rocks/csv/child_aggression.csv}
}
\usage{
child_aggression
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A study was carried out to explore the relationship between aggression and several potential predicting
factors in 666 children who had an older sibling. Variables measured were \strong{parenting_style} (high score = bad parenting practices),
\strong{computer_games} (high score = more time spent playing computer games), \strong{television} (high score = more time spent watching television),
\strong{diet} (high score = the child has a good diet low in harmful additives), and \strong{sibling_aggression} (high score = more aggression seen in their older sibling).
Past research indicated that parenting style and sibling aggression were good predictors of the level of aggression in the younger child.
The data contain the following variables:
\itemize{
\item \strong{aggression}: The child's aggression
\item \strong{television}: Time spent watching television (high score = more time spent watching television)
\item \strong{computer_games}: Time spent playing video games (high score = more time spent playing video games)
\item \strong{sibling_aggression}: Agression in older sibling (high score = more aggression seen in their older sibling).
\item \strong{diet}: The child's diet (high score = the child has a good diet low in harmful additives).
\item \strong{parenting_style}: the parent's parenting style (high score = bad parenting practices).
}
}
\keyword{datasets}
