% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticFun.R
\name{chiSQTest}
\alias{chiSQTest}
\title{Chi Square Test}
\usage{
chiSQTest(inp, ...)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
analyzes goodness of fit of fittet line to dataset
}
\description{
Chi Square Test
}

