% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{colorTransla}
\alias{colorTransla}
\title{function to convert a value to color}
\usage{
colorTransla(input, bigger = TRUE, minp = 0.05)
}
\arguments{
\item{input}{is the specific method output}

\item{bigger}{indicates if the "wanted" value is > or < 0.05}
}
\value{
returns the color name
}
\description{
function to convert a value to color
}

