% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{condom}
\alias{condom}
\title{Condom use data}
\format{
A tibble with 100 rows and 8 variables.
}
\source{
\href{https://www.discovr.rocks/csv/condom.csv}{www.discovr.rocks/csv/condom.csv}
}
\usage{
condom
}
\description{
A fictional dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about condom use. An HIV researcher explored the factors that influenced condom use with a new partner
(relationship less than 1 month old). The outcome measure was whether a condom was used. The predictor variables were mainly
scales from the Condom Attitude Scale (Sacco et al., 1991): gender; the degree to which the person views their relationship as 'safe'
from sexually transmitted disease; the degree to which previous experience influences attitudes towards condom use;
whether or not the couple used a condom in their previous encounter; the degree of self-control that a person has when it comes to condom use;
the degree to which the person perceives a risk from unprotected sex.
The data contains the following variables:
\itemize{
\item \strong{id}: Participant's ID
\item \strong{safety}: the degree to which the person views their relationship as 'safe' from sexually transmitted disease
\item \strong{use}: whether a condom was used (unprotected or condom used)
\item \strong{gender}: participant's gender (male or female)
\item \strong{experience}: the degree to which previous experience influences attitudes towards condom use
\item \strong{previous}: whether the couple used a condom in their previous encounter (condom used, no condom, this was the first time with this partner)
\item \strong{self_control}: the degree of self-control that a person has when it comes to condom use
\item \strong{risk_perception}: the degree to which the person perceives a risk from unprotected sex
}
}
\references{
\itemize{
\item Sacco, W. P., Levine, B., Reed, D. & Thompson, K. (1991). Attitudes about condom use as an AIDS-relevant behavior: Their factor structure and relation to condom use. \emph{Psychological Assessment: A Journal of Consulting and Clinical Psychology}, 3(2), 265–272. \doi{10.1037/1040-3590.3.2.265}
}
}
\keyword{datasets}
