% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{dark_lord}
\alias{dark_lord}
\title{Subliminal messages data}
\format{
A tibble with 64 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/dark_lord.csv}{www.discovr.rocks/csv/dark_lord.csv}
}
\usage{
dark_lord
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Both Ozzy Osbourne and Judas Priest have been accused of putting backward masked messages on their albums
that subliminally influence poor unsuspecting teenagers into doing things like blowing their heads off with
shotguns. A psychologist was interested in whether backward masked messages could have an effect. He created
a version of Britney Spears’ ‘Baby one more time’ that contained the masked message ‘deliver your soul to the
dark lord’ repeated in the chorus. He took this version, and the original, and played one version (randomly)
to a group of 32 people. Six months later he played them whatever version they hadn’t heard the time before.
So, each person heard both the original and the version with the masked message, but at different points in time.
The psychologist measured the number of satanic intrusions the person had in the week after listening to each
version. The (fictional) data contains the following variables:
\itemize{
\item \strong{id}: the participant's id
\item \strong{message}: whether the song had a subliminal satanic message or not
\item \strong{intrusions}: number of satanic intrusions in the week after hearing the song
}
}
\keyword{datasets}
