% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{dataAdjust}
\alias{dataAdjust}
\title{creates data.frame to htmlwidget json}
\usage{
dataAdjust(a)
}
\arguments{
\item{a}{is data matrix}
}
\value{
returns json list
}
\description{
creates data.frame to htmlwidget json
}

