% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{df_beta}
\alias{df_beta}
\title{DF beta data}
\format{
A tibble with 30 rows and 3 variables.
}
\source{
\href{https://www.discoveringstatistics.com/books/discovering-statistics-using-r/}{https://www.discoveringstatistics.com/books/discovering-statistics-using-r}
}
\usage{
df_beta
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictitious data to illustrate the DF Beta statistic. The tibble contains the following variables:
\itemize{
\item \strong{case}: a number from 0-30 indicating the entity (case)
\item \strong{x}: Imaginatively named predictor variable
\item \strong{y}: The creativity was flowing the day I generated these data - another imaginatively named variable. This time it's the outcome variable.
}
}
\keyword{datasets}
