% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{disc.data}
\alias{disc.data}
\title{output of numerical values}
\usage{
disc.data(x, method = "unPaired", preset = NULL)
}
\arguments{
\item{x}{is the dataframe input}

\item{method}{is "paired" or "unPaired"}

\item{preset}{is the invididual choosen preset ("facs","frap")}
}
\value{
returns a list of data.frames specific for the choosen method
}
\description{
output of numerical values
}

