% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{disc}
\alias{disc}
\title{discovr() main code}
\usage{
disc(x, y)
}
\arguments{
\item{x}{A data.frame or data.table}

\item{y}{Additional options}
}
\value{
The calculation output
}
\description{
discovr() main code
}

