% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{display}
\alias{display}
\title{Data about display rule understanding}
\format{
A tibble with 70 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/display.csv}{www.discovr.rocks/csv/display.csv}
}
\usage{
display
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about the relationship between theory of mind and display rule understanding. A display rule refers to displaying an appropriate emotion in a situation.
For example, if you receive a present that you don't like, you should smile politely and say "Thank you, Auntie Kate, I've always wanted a rotting cabbage"; you do not start
crying and scream "Why did you buy me a rotting cabbage?!" A psychologist believed that display rule understanding would be connecting to understanding others' mental states (having a 'theory of mind').
He measured children's understanding of display rules (with a task that they could pass or fail),
their age (months), and their ability to understand others' mental states ('theory of mind', measured with a false belief task that they could pass or fail).
The data contains the following variables:
\itemize{
\item \strong{id}: Participant's ID
\item \strong{age}: the participant's age in months
\item \strong{fb}: Whether someone passed the false belief task (yes or no)
\item \strong{display}: Whether someone passed the display rule understanding task (yes or no)
}
}
\keyword{datasets}
