% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{distracted_by_social_media}
\alias{distracted_by_social_media}
\title{Data about being distracted by social media}
\format{
A tibble with 8 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/distracted_by_social_media.csv}{www.discovr.rocks/csv/distracted_by_social_media.csv}
}
\usage{
distracted_by_social_media
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about being distracted by social media. On my statistics module students have weekly practical statistics
classes in a computer laboratory. I've noticed that many students are studying social media more than the very interesting
statistics assignments that I have set them. I wanted to see the impact that this behaviour had on their exam performance.
I collected data from all 260 students on my module. I classified their attendance as being either more or less than 50\% of their practical classes,
and I classified them as someone who I'd noticed sneakily looking at social_media during their class, or someone who I'd never seen looking at social media.
After the exam, I noted whether they passed or failed.
The data contains the following variables:
\itemize{
\item \strong{attendance}: Whether someone was classified as attending more or less than 50\% of their practical classes
\item \strong{social_media}: Whether someone was classified as a person who looked a social media during class (or not)
\item \strong{exam}: whether someone passed or failed their statistics exam
\item \strong{frequency}: the number of people falling into each combination of attendance, social media and exam performance
}
}
\keyword{datasets}
