% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{dog_training}
\alias{dog_training}
\title{Dog training data}
\format{
A tibble with 668 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/dog_training.csv}{www.discovr.rocks/csv/dog_training.csv}
}
\usage{
dog_training
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about dogs being trained to vocalize whenever they sniff an alien life form. Essentially dogs were trained using food rewards. One each trial they sniffed an alien and if they made a vocalization they were rewarded with food.
This data shows how vocalisations change over blocks of these training trials. The tibble contains the following variables:
\itemize{
\item \strong{id}: name of the fictional dog. (Fun fact, the names are real pet names randomly selected from the pet registry in Seattle: \url{https://randommer.io/pet-names})
\item \strong{trials}: the block of trials (each block represents 100 trials, so block 1 is the result of the first 100 trials and 5 is the result of trials 400-500)
\item \strong{vocalizations}: the percentage of trials during which the dog vocalised.
}
}
\keyword{datasets}
