% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{download}
\alias{download}
\title{Download festival data}
\format{
A tibble with 810 rows and 5 variables.
}
\source{
\href{https://www.discovr.rocks/csv/download.csv}{www.discovr.rocks/csv/download.csv}
}
\usage{
download
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about people stinking at music festivals. A biologist was worried about the potential health effects of
music festivals. She went to the Download Music Festival and measured the hygiene of 810 concert-goers over the three days of the festival.
She tried to measure every person on every day but, because it was difficult to track people down, there were missing data on days 2 and 3.
Hygiene was measured using a standardized technique that results in a score ranging between 0 (you smell like a corpse that’s been left to rot up a skunk’s arse)
and 4 (you smell of sweet roses on a fresh spring day). I know from bitter experience that sanitation is not always great at these places and so the biologist
predicted that personal hygiene would go down dramatically over the three days of the festival. The object contains the following variables:
\itemize{
\item \strong{ticket_no}: the ticket number of the participant as a factor
\item \strong{gender}: The gender with which the participant self-identifies as a factor (male, female, non-binary)
\item \strong{day1}: the hygiene score from 0 (eau de toilet) to 4 (eau de toilette) on day 1 of the festival
\item \strong{day2}: the hygiene score from 0 (eau de toilet) to 4 (eau de toilette) on day 2 of the festival
\item \strong{day3}: the hygiene score from 0 (eau de toilet) to 4 (eau de toilette) on day 3 of the festival
}
}
\keyword{datasets}
