% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{eddiefy}
\alias{eddiefy}
\title{Iron Maiden Spotify song features data}
\format{
A tibble with 173 rows and 18 variables.
}
\source{
\href{https://www.discovr.rocks/csv/eddiefy.csv}{www.discovr.rocks/csv/eddiefy.csv}
}
\usage{
eddiefy
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A dataset containing the song features data from the \href{https://developer.spotify.com/documentation/web-api/}{Spotify API} for the studio albums (190-2015) of the greatest band ever, \href{www.ironmaiden.com}{Iron Maiden}.
Data were obtained using the \href{https://www.rcharlie.com/spotifyr/}{spotifyr} package.
\itemize{
\item \strong{artist_name}: Pointless variable that reminds us that the data relate to Iron Maiden
\item \strong{album_name}: Name of the album
\item \strong{track_name}: Name of the song
\item \strong{year}: Year of release of the album
\item \strong{danceability}: From the Spotify API: "Danceability describes how suitable a track is for dancing based on a combination of musical elements including tempo, rhythm stability, beat strength, and overall regularity. A value of 0.0 is least danceable and 1.0 is most danceable."
\item \strong{energy}: From the Spotify API: "Energy is a measure from 0.0 to 1.0 and represents a perceptual measure of intensity and activity. Typically, energetic tracks feel fast, loud, and noisy. For example, death metal has high energy, while a Bach prelude scores low on the scale. Perceptual features contributing to this attribute include dynamic range, perceived loudness, timbre, onset rate, and general entropy."
\item \strong{key}: From the Spotify API: "The key the track is in. Integers map to pitches using standard Pitch Class notation . E.g. 0 = C, 1 = C-sharp/D-flat, 2 = D, and so on."
\item \strong{loudness}: From the Spotify API: "The overall loudness of a track in decibels (dB). Loudness values are averaged across the entire track and are useful for comparing relative loudness of tracks. Loudness is the quality of a sound that is the primary psychological correlate of physical strength (amplitude). Values typical range between -60 and 0 db."
\item \strong{mode}: From the Spotify API: "Mode indicates the modality (major or minor) of a track, the type of scale from which its melodic content is derived. Major is represented by 1 and minor is 0."
\item \strong{speechiness}: From the Spotify API: "Speechiness detects the presence of spoken words in a track. The more exclusively speech-like the recording (e.g. talk show, audio book, poetry), the closer to 1.0 the attribute value. Values above 0.66 describe tracks that are probably made entirely of spoken words. Values between 0.33 and 0.66 describe tracks that may contain both music and speech, either in sections or layered, including such cases as rap music. Values below 0.33 most likely represent music and other non-speech-like tracks."
\item \strong{acousticness}: From the Spotify API: "A confidence measure from 0.0 to 1.0 of whether the track is acoustic. 1.0 represents high confidence the track is acoustic."
\item \strong{instrumentalness}: From the Spotify API: "Predicts whether a track contains no vocals. “Ooh” and “aah” sounds are treated as instrumental in this context. Rap or spoken word tracks are clearly “vocal”. The closer the instrumentalness value is to 1.0, the greater likelihood the track contains no vocal content. Values above 0.5 are intended to represent instrumental tracks, but confidence is higher as the value approaches 1.0."
\item \strong{liveness}: From the Spotify API: "Detects the presence of an audience in the recording. Higher liveness values represent an increased probability that the track was performed live. A value above 0.8 provides strong likelihood that the track is live."
\item \strong{valence}: From the Spotify API: "A measure from 0.0 to 1.0 describing the musical positiveness conveyed by a track. Tracks with high valence sound more positive (e.g. happy, cheerful, euphoric), while tracks with low valence sound more negative (e.g. sad, depressed, angry)."
\item \strong{tempo}: From the Spotify API: "The overall estimated tempo of a track in beats per minute (BPM). In musical terminology, tempo is the speed or pace of a given piece and derives directly from the average beat duration."
\item \strong{time_signature}: From the Spotify API: "An estimated overall time signature of a track. The time signature (meter) is a notational convention to specify how many beats are in each bar (or measure)."
\item \strong{song_duration}: Song length as a \code{lubridate} duration object
\item \strong{duration_ms}: Song length in milliseconds as an integer value.
}
}
\keyword{datasets}
