% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{elephooty}
\alias{elephooty}
\title{Elephant football data}
\format{
A tibble with 120 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/elephooty.csv}{www.discovr.rocks/csv/elephooty.csv}
}
\usage{
elephooty
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about elephant football. The highlight of the elephant calendar is the annual elephant soccer event
in Nepal. A heated argument burns between the African and Asian elephants. In 2010, the president of the
Asian Elephant Football Association, an elephant named Boji, claimed that Asian elephants were more talented
than their African counterparts. The head of the African Elephant Soccer Association, an elephant called Tunc,
issued a press statement that read 'I make it a matter of personal pride never to take seriously any remark made
by something that looks like an enormous scrotum'. I was called in to settle things.
I collected data from the two types of elephants (Asian or African) over a season and recorded how many goals
each elephant scored and how many years of experience the elephant had. The data set has four variables:
\itemize{
\item \strong{id}: Elephant id
\item \strong{elephant}: Whether the elephant was an Asian elephant or an African elephant
\item \strong{experience}: how many years of football experience the elephant had
\item \strong{goals}: how many goals the elephant scored during the season
}
}
\keyword{datasets}
