% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{essay_marks}
\alias{essay_marks}
\title{Essay mark data}
\format{
A tibble with 45 rows and 4 variables.
}
\source{
\url{https://www.discovr.rocks}
}
\usage{
essay_marks
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about essay marks. A student was interested in whether there was a positive relationship between the time spent doing an essay and the mark received.
He got 45 of his friends and timed how long they spent writing an essay (hours) and the percentage they got in the essay (essay).
He also translated these grades into their degree classifications (grade): in the UK, a student can get a first-class mark (the best), an upper-second-class mark,
a lower second, a third, a pass or a fail (the worst).
The data set has four variables
\itemize{
\item \strong{id}: Student id
\item \strong{essay}: Percentage mark on the essay
\item \strong{hours}: hours spend writing the essay
\item \strong{grade}: factor that converts the essay percentage to the degree classification of the essay (see general description)
}
}
\keyword{datasets}
