% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{field_2006}
\alias{field_2006}
\title{Field (2006) data}
\format{
A tibble with 381 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/field_2006.csv}{www.discovr.rocks/csv/gallup_2003.csv}
}
\usage{
field_2006
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Early in my career I looked at the effect of giving children information about entities. In one study (Field, 2006),
I used three novel entities (the quoll, quokka and cuscus) and children were told threat information about one of the entities,
positive information about another, and given no information about the third (our control). After the information I asked the
children to place their hands in three wooden boxes each of which they believed contained one of the aforementioned entities
The data from the study has three variables:
\itemize{
\item \strong{id}: The participant's id (these do not come from the study data file)
\item \strong{info_type}: the type of information given about the animal
\item \strong{latency}: the time taken for the child to approach the box (children who had not approached the box within 15s were assumed to be not consenting to that task and were scored as 15s)
}
}
\references{
Field, A. P. (2006). The behavioral inhibition system and the verbal information pathway to children's fears. \emph{Journal of Abnormal Psychology}, \emph{115}, 742–752. \doi{10.1037/0021-843x.115.4.742}
}
\keyword{datasets}
