% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticFun.R
\name{fTest}
\alias{fTest}
\title{F-Test}
\usage{
fTest(input)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
compares variances of the samples - only normal distribution
}
\description{
F-Test
}

