% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{gallup_2003}
\alias{gallup_2003}
\title{Gelman & Weakliem (2009) data}
\format{
A tibble with 15 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/gallup_2003.csv}{www.discovr.rocks/csv/gallup_2003.csv}
}
\usage{
gallup_2003
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
It's something of a wonder how evolution managed to produce such a monstrosity as the human penis.
One theory is sperm competition: the human penis has an unusually large glans (the 'bell-end') compared to other primates,
and this may have evolved so that the penis can displace seminal fluid from other males by 'scooping it out' during intercourse.
Armed with various devices from Hollywood Exotic Novelties, an artificial vagina from California Exotic Novelties, and some water and cornstarch
Gallup et al. (2003) put this theory to the test. They loaded the artificial vagina with 2.6 ml of fake sperm and inserted one
of three female sex toys into it before withdrawing it: a control phallus that had no coronal ridge (i.e., no bell-end),
a phallus with a minimal coronal ridge (small bell-end) and a phallus with a coronal ridge. They measured sperm displacement as a percentage:
100\% means that all the sperm was displaced, and 0\% means that none of the sperm was displaced. If the human penis evolved as a sperm displacement
device then Gallup et al. predicted: (1) that having a bell-end would displace more sperm than not; and
(2) that the phallus with the larger coronal ridge would displace more sperm than the phallus with the minimal coronal ridge. The data from the study has three variables:
\itemize{
\item \strong{id}: The participant's id (these do not come from the study data file)
\item \strong{phallus}: the type of phallus used (No coronal ridge, minimal coronal ridge and coronal ridge)
\item \strong{displace}: percentage of sperm displaced by the phallus
}
}
\references{
\itemize{
\item Gallup, G. G. J., Burch, R. L., Zappieri, M. L., Parvez, R., Stockwell, M., & Davis, J. A. (2003). The human penis as a semen displacement device. \emph{Evolution and Human Behavior}, \emph{24}, 277–289. \href{https://doi.org/10.1016/S1090-5138(03)00016-3}{doi.org/10.1016/S1090-5138(03)00016-3}
}
}
\keyword{datasets}
