% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{gelman_2009}
\alias{gelman_2009}
\title{Gelman & Weakliem (2009) data}
\format{
A tibble with 548 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/gelman_2009.csv}{www.discovr.rocks/csv/gelman_2009.csv}
}
\usage{
gelman_2009
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Apparently there are more beautiful women in the world than there are handsome men. Satoshi Kanazawa explains this finding
in terms of good-looking parents being more likely to have a baby daughter as their first child than a baby son.
Perhaps more controversially, he suggests that, from an evolutionarily perspective, beauty is a more valuable trait for women than
for men (Kanazawa, 2007). In a playful and very informative paper, Andrew Gelman and David Weakliem discuss various
statistical errors and misunderstandings, some of which have implications for Kanazawa's claims.
The 'playful' part of the paper is that to illustrate their point they collected data on the 50 most beautiful celebrities
(as listed by People magazine) of 1995-2000. They counted how many male and female children they had as of 2007.
If Kanazawa is correct, these beautiful people would have produced more girls than boys.
These are the data from that study. The data contains the following variables:
\itemize{
\item \strong{person}: The name of the celebrity
\item \strong{child}: whether children are sons or daughters
\item \strong{number}: the number of sons/daughters (depending on the value of child) the celebrity has (at the time of the study)
}
}
\references{
\itemize{
\item Gelman, A., & Weakliem, D. (2009). Of beauty, sex and power: Too little attention has been paid to the statistical challenges in estimating small effects. \emph{American Scientist}, 97, 310–316. \url{https://www.jstor.org/stable/27859361}
}
}
\keyword{datasets}
