% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{glastonbury}
\alias{glastonbury}
\title{Glastonbury festival data}
\format{
A tibble with 810 rows and 5 variables.
}
\source{
\href{https://www.discovr.rocks/csv/glastonbury.csv}{www.discovr.rocks/csv/glastonbury.csv}
}
\usage{
glastonbury
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
More fictional data about people stinking at music festivals. The same biologist who was worried about the potential health effects of music festivals and
collected data at a heavy metal festival (Download Festival), was worried that her findings might not generalize. To find out whether the type of music a person
likes predicts whether hygiene decreases over the festival the biologist measured hygiene over the three days of the Glastonbury Music Festival, which has an eclectic clientele.
Her hygiene measure ranged between 0 (you smell like you've bathed in sewage) and 4 (you smell like you've bathed in freshly baked bread).
The biologist coded the festival-goer's musical affiliations into the categories 'hipster' (people who mainly like alternative music), 'metalhead' (people who like heavy metal), and 'raver' (people who like dance/ambient stuff).
Anyone not falling into these categories was labelled 'no subculture'. The object contains the following variables:
\itemize{
\item \strong{ticket_no}: the ticket number of the participant as a factor
\item \strong{subculture}: The musical subculture with which the participant self-identifies as a factor (no subculture, hipster, metalhead, raver)
\item \strong{day1}: the hygiene score from 0 (eau de toilet) to 4 (eau de toilette) on day 1 of the festival
\item \strong{day2}: the hygiene score from 0 (eau de toilet) to 4 (eau de toilette) on day 2 of the festival
\item \strong{day3}: the hygiene score from 0 (eau de toilet) to 4 (eau de toilette) on day 3 of the festival
\item \strong{change}: the change in hygiene score from day 1 to day 3 of the festival
}
}
\keyword{datasets}
