% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{goggles}
\alias{goggles}
\title{Beer goggles effect data}
\format{
A tibble with 48 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/goggles.csv}{www.discovr.rocks/csv/goggles.csv}
}
\usage{
goggles
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about the beer goggles effect. An anthropologist was interested in the effects of facial
attractiveness on the beer-goggles effect. She randomly selected 48 participants. Participants were randomly
subdivided into three groups of 16: (1) a placebo group drank 500 ml of alcohol-free beer; (2) a low-dose group drank
500 ml of average strength beer (4\% ABV); and (3) a high-dose group drank 500 ml of strong beer (7\% ABV).
Within each group, half (\emph{n} = 8) rated the attractiveness of 50 photos of unattractive faces on a scale from
0 (pass me a paper bag) to 10 (pass me their phone number) and the remaining half rated 50 photos of attractive faces.
The outcome for each participant was their median rating across the 50 photos. The data set has four variables
\itemize{
\item \strong{id}: Participant's id
\item \strong{facetype}: Whether the participant rated photos of 'attractive' or 'unattractive' faces
\item \strong{alcohol}: The alcohol group to which the participant was assigned. Either a placebo group (who drank 500 ml of alcohol-free beer), a low-dose group
(who drank 500 ml of 4\% ABV beer), or a high-dose group (who drank 500 ml of 7\% ABV beer)
\item \strong{attractiveness}: the median rating of the attractiveness  of 50 photos from 0 (pass me a paper bag) to 10 (pass me their phone number)
}
}
\keyword{datasets}
