% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{goggles_lighting}
\alias{goggles_lighting}
\title{Beer goggles and lighting data}
\format{
A tibble with 208 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/goggles_lighting.csv}{www.discovr.rocks/csv/goggles.csv}
}
\usage{
goggles_lighting
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about the moderating effect of lighting on the beer goggles effect. In previous edxample we came across the beer-goggles
which suggests that alcohol impairs judgements of facial attractiveness. In this fictional follow-up study a sample of 26 people
are given doses of alcohol (0 pints, 2 pints, 4 pints and 6 pints of lager) over four different weeks. They are asked to rate a
bunch of photos of faces in either dim or bright lighting. The outcome measure was the mean attractiveness rating (out of 100)
of the faces and the predictors were the dose of alcohol and the lighting conditions  The data set has four variables
\itemize{
\item \strong{id}: Participant's id
\item \strong{lighting}: Whether the photos were viewed in dim or bright lighting
\item \strong{alcohol}: The dose of alcohol taken before ratings were made
\item \strong{rating}: the median rating of the attractiveness  of the photos rated from 0 (pass me a paper bag) to 10 (pass me their phone number)
}
}
\keyword{datasets}
