% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{grades}
\alias{grades}
\title{Grades data}
\format{
A tibble with 25 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/grades.csv}{www.discovr.rocks/csv/grades.csv}
}
\usage{
grades
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about stats grades. As a statistics lecturer I am interested in the factors that determine whether a
student will do well on a statistics course. Imagine I took 25 students and looked at their grades for my statistics
module at the end of their first year at university:  first class, upper second class, lower second class, third class, pass and fail.
I also asked these students what grade they got in their high school maths exams. In the UK GCSEs are school exams taken at age 16
that are graded A, B, C, D, E or F (an A grade is the best). The data set has three variables
\itemize{
\item \strong{id}: Student id
\item \strong{stats}: Degree classification for a statistics module
\item \strong{gcse}: GCSE mathematics classification at age 16
}
}
\keyword{datasets}
