% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{graphicGen}
\alias{graphicGen}
\title{splitts calculation in paired and unpaired sections}
\usage{
graphicGen(x, method, colNames, dataSetName)
}
\arguments{
\item{x}{is the list of data matrixes}

\item{method}{is the method name for the widget}

\item{colNames}{contains all column names for the widget}
}
\value{
returns the final input for the disc.graphics module
}
\description{
splitts calculation in paired and unpaired sections
}

