% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{handlebars}
\alias{handlebars}
\title{Data about riding bikes}
\format{
A tibble with 4 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/handlebars.csv}{www.discovr.rocks/csv/handlebars.csv}
}
\usage{
handlebars
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about national trends in bike riding. I have a soft spot for the Netherlands.
The Dutch travel by bike much more than the English. I noticed that many more Dutch people cycle while steering with only one hand.
I pointed this out to one of my Dutch friends, Birgit Mayer, and she said that I was a crazy English fool and that Dutch people did not cycle one-handed.
Several weeks of me pointing at one-handed cyclists and her pointing at two-handed cyclists ensued. To put it to the test
I counted the number of Dutch and English cyclists who ride with one or two hands on the handlebars. I didn't really do this.
The data contains the following variables:
\itemize{
\item \strong{hands}: Whether someone was rode a bike with one or two hands on the handlebars
\item \strong{nationality}: Whether someone was Dutch or English
\item \strong{frequency}: the number of people falling into each combination of hands and nationality
}
}
\keyword{datasets}
