% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{hangover}
\alias{hangover}
\title{Hangover cure data}
\format{
A tibble with 15 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/hangover.csv}{www.discovr.rocks/csv/hangover.csv}
}
\usage{
hangover
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A marketing manager tested the benefit of soft drinks for curing hangovers. He took 15 people and got
them drunk. The next morning as they awoke, dehydrated and feeling as though they’d licked a camel’s
sandy feet clean with their tongue, he gave five of them water to drink, five of them Lucozade
(a very nice glucose-based UK drink) and the remaining five a leading brand of cola. He measured how well they felt
(on a scale from 0 = I feel like death to 10 = I feel really full of beans and healthy) two hours later.
He measured how drunk the person got the night before on a scale of 0 = as sober as a nun to 10 = flapping about
like a haddock out of water on the floor in a puddle of their own vomit. These data are fictional.
The object contains the following variables:
\itemize{
\item \strong{id}: participant id
\item \strong{drink}: whether the person drank water, Lucozade or Cola as a hangover cure
\item \strong{well}: how well the person felt two hours after the hangover cure (0 = I feel like death to 10 = I feel really full of beans and healthy)
\item \strong{drunk}: how drunk the person got the night before (0 = as sober as a nun to 10 = flapping about like a haddock out of water on the floor in a puddle of their own vomit)
}
}
\keyword{datasets}
