% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{hiccups}
\alias{hiccups}
\title{Hiccups data}
\format{A tibble with 60 rows and 3 variables.}
\source{
\url{https://www.discoveringstatistics.com/books/discovering-statistics-using-r/}
}
\usage{
hiccups
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
People have many methods for stopping hiccups (a surprise, holding your breath), and medical science has
put its collective mind to the task too. The official treatment methods include tongue-pulling manoeuvres,
massage of the carotid artery, and, believe it or not, digital rectal massage (Fesmire, 1988).
Let’s say we wanted to put digital rectal massage to the test (erm, as a cure for hiccups).
We took 15 hiccup sufferers, and during a bout of hiccups administered each of the three procedures
(in random order and at intervals of 5 minutes) after taking a baseline of how many hiccups they had per minute.
We counted the number of hiccups in the minute after each procedure. These data are fictional.
The object contains the following variables:

\itemize{
\item \strong{id}: participant id
\item \strong{intervention}: the 4 interventions that each participant tried
\item \strong{hiccups}: the number of hiccups during the minute after the intervention

}
}
\keyword{datasets}
