% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{honesty_lab}
\alias{honesty_lab}
\title{Honesty lab data}
\format{
A tibble with 100 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/honesty_lab.csv}{www.discovr.rocks/csv/honesty_lab.csv}
}
\usage{
honesty_lab
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about the honesty lab. Imagine we were interested in how people evaluated dishonest acts.
Participants evaluate the dishonesty of acts based on watching videos of people confessing to those acts.
Imagine we took 100 people and showed them a random dishonest act described by the perpetrator.
They then evaluated the honesty of the act (from 0 = appalling behaviour to 10 = it’s OK really)
and how much they liked the person (0 = not at all, 10 = a lot). The data set has three variables
\itemize{
\item \strong{id}: Participant's id
\item \strong{deed}: evaluation of the honesty of the act (from 0 = appalling behaviour to 10 = it’s OK really)
\item \strong{likeableness}: evaluation of the perpetrator (0 = not at all, 10 = a lot)
}
}
\keyword{datasets}
