% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{horoscope}
\alias{horoscope}
\title{Data about horoscopes}
\format{
A tibble with 48 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/horoscope.csv}{www.discovr.rocks/csv/horoscope.csv}
}
\usage{
horoscope
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data exploring whether horoscopes are tomfoolery. In a hypothetical universe, I recruited 2201 people, made a note of their star sign
(this variable, obviously, has 12 categories: Capricorn, Aquarius, Pisces, Aries, Taurus, Gemini, Cancer, Leo, Virgo, Libra, Scorpio and Sagittarius)
and whether they believed in horoscopes (this variable has two categories: believer or unbeliever). I sent them an identical horoscope about events
in the next month, which read "August is an exciting month for you. You will make friends with a tramp in the first week and cook him a cheese omelette.
Curiosity is your greatest virtue, and in the second week, you'll discover knowledge of a subject that you previously thought was boring. Statistics perhaps.
You might purchase a book around this time that guides you towards this knowledge. Your new wisdom leads to a change in career around the third week,
when you ditch your current job and become an accountant. By the final week you find yourself free from the constraints of having friends,
your boy/girlfriend has left you for a Russian ballet dancer with a glass eye, and you now spend your weekends doing loglinear analysis
by hand with a pigeon called Hephzibah for company." At the end of August I interviewed these people and I classified the horoscope as having come true, or not,
based on how closely their lives had matched the fictitious horoscope. The data contains the following variables:
\itemize{
\item \strong{star_sign}: The person's star sign (Capricorn, Aquarius, Pisces, Aries, Taurus, Gemini, Cancer, Leo, Virgo, Libra, Scorpio and Sagittarius)
\item \strong{believe}: Whether the perwson believes in horoscopes (unbeliever or believer)
\item \strong{true}: Whether the horoscope came true (Horoscope didn't come true, Horoscope came true)
\item \strong{frequency}: the number of people falling into each combination of star sign, belief in star signs and whether the horoscope came true
}
}
\keyword{datasets}
