% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{invisibility_base}
\alias{invisibility_base}
\title{Cloak of invisibility data (pre-post design)}
\format{
A tibble with 80 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/invisibility_base.csv}{www.discovr.rocks/csv/invisibility_base.csv}
}
\usage{
invisibility_base
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
In \link{invisibility_cloak} we compared the number of mischievous acts in people who had invisibility cloaks to
those without. Imagine we replicated that study, but changed the design so that we recorded the number of
mischievous acts in these participants before the study began as well as during the study.
The data contains the following variables:
\itemize{
\item \strong{id}: participant id
\item \strong{cloak}: whether the participant was assigned a cloak of invisibility
\item \strong{mischief_pre}: the number of mischievous acts committed during the week before the study
\item \strong{mischief}: the number of mischievous acts committed during the week of the study
}
}
\keyword{datasets}
