% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{invisibility_cloak}
\alias{invisibility_cloak}
\title{Cloak of invisibility data (independent design)}
\format{
A tibble with 24 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/invisibility.csv}{www.discovr.rocks/csv/invisibility.csv}
}
\usage{
invisibility_cloak
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
I got very excited by two news stories implying that scientists had made Harry Potter's cloak of invisibility. Although the newspapers overstated the case,
I imagined a future in which we have cloaks of invisibility to test out. Given my slightly mischievous streak, the future me is interested in the effect
that wearing a cloak of invisibility has on the tendency for mischief. I take 24 participants and place them in an enclosed community.
The community is riddled with hidden cameras so that we can record mischievous acts. Half of the participants are given cloaks of invisibility;
they are told not to tell anyone else about their cloak and that they can wear it whenever they liked. I measure how many mischievous acts they performed in one week.
The object contains the following variables:
\itemize{
\item \strong{id}: participant id
\item \strong{cloak}: whether the participant was assigned a cloak of invisibility
\item \strong{mischief}: the number of mischievous acts committed during a week
}
}
\keyword{datasets}
