% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{invisibility_rm}
\alias{invisibility_rm}
\title{Cloak of invisibility data (repeated measures design)}
\format{
A tibble with 24 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/invisibility_rm.csv}{www.discovr.rocks/csv/invisibility_rm.csv}
}
\usage{
invisibility_rm
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
I got very excited by two news stories implying that scientists had made Harry Potter’s cloak of invisibility. Although the newspapers overstated the case,
I imagined a future in which we have cloaks of invisibility to test out. Given my slightly mischievous streak, the future me is interested in the effect
that wearing a cloak of invisibility has on the tendency for mischief. I take 12 participants and place them in an enclosed community.
The community is riddled with hidden cameras so that we can record mischievous acts. For one week the participants are given cloaks of invisibility, during a different week they are not.
I measure how many mischievous acts they performed in each week. These data are the same as in \link{invisibility_cloak} but arranged in a repeated measures design.
The object contains the following variables:
\itemize{
\item \strong{id}: participant id
\item \strong{cloak}: whether the participant had access to a cloak of invisibility
\item \strong{mischief}: the number of mischievous acts committed during a week
}
}
\keyword{datasets}
