% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{lacourse_2001_females}
\alias{lacourse_2001_females}
\title{Lacourse et al. (2001) data}
\format{
A tibble with 121 rows and 15 variables.
}
\source{
\href{https://www.discovr.rocks/csv/lacourse_2001_females.csv}{www.discovr.rocks/csv/lacourse_2001_females.csv}
}
\usage{
lacourse_2001_females
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Eric Lacourse and his colleagues used questionnaires to measure suicide risk (yes or no), marital status of parents (together or divorced/separated), the extent to which the person experienced neglect,
self-estrangement/powerlessness (adolescents who have negative self-perceptions, are bored with life, etc.), social isolation (feelings of a lack of support), normlessness (beliefs that socially disapproved behaviours can be used to achieve certain goals),
meaninglessness (doubting that school is relevant to gaining employment) and drug use. In addition, the authors measured liking of heavy metal; they included the sub-genres of classic (Black Sabbath, Iron Maiden), thrash metal (Slayer, Metallica),
death/black metal (Obituary, Burzum) and gothic (Marilyn Manson). As well as liking they measured behavioural manifestations of worshipping these bands (e.g., hanging posters, hanging out with other metal fans) and what the authors termed 'vicarious music listening'
(whether music was used when angry or to bring out aggressive moods). They used logistic regression to predict suicide risk from these variables for those identifying as male and female separately. This the data for the female sample and contains the following variables:
\itemize{
\item \strong{age}: Participant's age
\item \strong{age_group}: Age group of participant (14-16 years old or 16-19 years old)
\item \strong{drug_use}: the quantity of alcohol, cannabis, and hallucinogens consumed during the past year from 1 (never) to 4 (10 or more times)
\item \strong{gender}: the self-reported gender (all female in this sub-sample)
\item \strong{isolation}: a measure of subjective perceptions of lack of support
\item \strong{marital_status}: whether the participants parents were together or separated/divorced
\item \strong{meaninglessness}: a measure of the degree to which the person doubts the relevance of school in attaining future employment
\item \strong{metal}: the degree to which the person likes heavy metal music
\item \strong{mother_negligence}: the perceived warmth in relationships with the mother. A high score represents a perception of cold and rejecting maternal relationship.
\item \strong{father_negligence}: the perceived warmth in relationships with the father A high score represents a perception of cold and rejecting paternal relationship.
\item \strong{normlessness}: a measure of beliefs that socially disapproved behaviours may be used to achieve certain goals
\item \strong{self_estrangement}: a measure of negative self-perception, feeling bored with life, overwhelmed by difficulties considered out of ones control.
\item \strong{suicide_risk}: whether the participant was at risk for suicide or not
\item \strong{vicarious}: the degree to which the person listens to music when angry or to bring out anger
\item \strong{worshipping}: the degree to which the person engages in fan-like behaviour about the bands they like
}
}
\references{
\itemize{
\item Lacourse, E., Claes, M., & Villeneuve, M. (2001). Heavy metal music and adolescent suicidal risk. \emph{Journal of Youth and Adolescence}, 30, 321–332. \doi{10.1023/A:1010492128537}
}
}
\keyword{datasets}
