% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{lambert_2012}
\alias{lambert_2012}
\title{Lambert et al. (2012) data}
\format{
A tibble with 240 rows and 6 variables
}
\source{
\href{https://www.discovr.rocks/csv/lambert_2012.csv}{www.discovr.rocks/csv/lambert_2012.csv}
}
\usage{
lambert_2012
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Lambert et al. (2012) found that pornography is related to infidelity. This object contains the data from that study.
\itemize{
\item \strong{id}: participant ID (not from the original data)
\item \strong{consumption}: pornography consumption on a scale from 0 (low) to 8 (high)
\item \strong{ln_porn}: log transformed values of consumption
\item \strong{commit}: commitment to the participant's current relationship on a scale from 1 (low) to 5 (high)
\item \strong{phys_inf}: whether the person had committed a physical act that they or their partner would consider to be unfaithful (0 = no, 1 = one of them would consider it unfaithful, 2 = both of them would consider it unfaithful)
\item \strong{hook_up}: the number of people they had 'hooked up' with in the previous year. (A 'hook-up' was defined to participants as 'when two people get together for a physical encounter and don't necessarily expect anything further')
}
}
\references{
\itemize{
\item Lambert, N. M., Negash, S., Stillman, T. F., Olmstead, S. B., & Fincham, F. D. (2012). A love that doesn't last: Pornography consumption and weakened commitment to one's romantic partner. \emph{Journal of Social and Clinical Psychology}, 31, 410–438. \doi{10.1521/jscp.2012.31.4.410}
}
}
\keyword{datasets}
