% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{listOut}
\alias{listOut}
\title{function to recover data from JSON list}
\usage{
listOut(input)
}
\arguments{
\item{input}{is the specific JSON list dataset}
}
\value{
returns the recovered data.frame
}
\description{
function to recover data from JSON list
}

