% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticFun.R
\name{mannWhitTest}
\alias{mannWhitTest}
\title{Mann-Whitney U test - Wilcoxon sum rank test}
\usage{
mannWhitTest(input)

mannWhitTest(input)
}
\arguments{
\item{input}{A data.frame or data.table}

\item{input}{A data.frame or data.table}
}
\value{
Alternative test for the paired t-test if data is not normal - on independent samples

shows if datadimensionality can be reduced
}
\description{
Mann-Whitney U test - Wilcoxon sum rank test

PCA dimension reduction
}

