% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{mcnulty_2008}
\alias{mcnulty_2008}
\title{McNulty et al. (2008) data}
\format{
A tibble with 164 rows and 5 variables
}
\source{
\href{https://www.discovr.rocks/csv/mcnulty_2008.csv}{www.discovr.rocks/csv/mcnulty_2008.csv}
}
\usage{
mcnulty_2008
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
McNulty et al. (2008) found a relationship between a person’s attractiveness and how much support they give their partner among newlywed heterosexual couples.
These data simulate the results of that study. The object contains the following variables:
\itemize{
\item \strong{id}: participant ID
\item \strong{attractiveness}: attractiveness of participant
\item \strong{support}: support given to partner
\item \strong{satisfaction}: relationship satisfaction
\item \strong{spouse}: whether the participant is a husband or wife
}
}
\references{
\itemize{
\item McNulty, J. K., Neff, L. A., & Karney, B. R. (2008). Beyond initial attraction: Physical attractiveness in newlywed marriage. \emph{Journal of Family Psychology}, 22, 135–143. \url{https://doi.org/10.1037/0893-3200.22.1.135}
}
}
\keyword{datasets}
