% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{men_dogs}
\alias{men_dogs}
\title{Are men like dogs data}
\format{
A tibble with 40 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/men_dogs.csv}{www.discovr.rocks/csv/men_dogs.csv}
}
\usage{
men_dogs
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A psychologist was interested in the cross-species differences between men and dogs. She observed a group
of dogs and a group of men in a naturalistic setting (20 of each). She classified several behaviours as being
dog-like (urinating against trees and lampposts, attempts to copulate with anything that moved, and attempts to
lick their own genitals). For each man and dog she counted the number of dog-like behaviours displayed in a
24-hour period.  The (fictional) data contains the following variables:
\itemize{
\item \strong{id}: the participant's id
\item \strong{species}: whether the participant was a man or a dog
\item \strong{behaviour}: number of dog-like behaviours exhibited by the participant in 24 hours
}
}
\keyword{datasets}
