% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{metal}
\alias{metal}
\title{Metal music and anger}
\format{
A tibble with 90 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/metal.csv}{www.discovr.rocks/csv/metal.csv}
}
\usage{
metal
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
People have claimed that listening to heavy metal, because of its aggressive sonic palette and often violent or emotionally negative lyrics,
leads to angry and aggressive behaviour. As a very non-violent metal fan this accusation bugs me
(BTW there are some real data on this in \link{sharman_2015}). Imagine I designed a study to test this possibility.
I took groups of self-classifying metalheads and non-metalheads (fan) and assigned them randomly to listen to 15 minutes of either the
sound of an angle grinder scraping a sheet of metal (control noise), metal music, or pop music (soundtrack).
Each person rated their anger on a scale ranging from 0 (\emph{All you need is love, da, da, da-da-da}) to 100 (\emph{---- me, I'm all out of enemies}).
These data are fictitious.
\itemize{
\item \strong{id}: the participant's ID
\item \strong{soundtrack}: whether the participant listened to 15 minutes of an angle grinder, metal music or pop music.
\item \strong{fan}: whether the participant self-classified as a metal fan (\emph{metalhead}) or not.
\item \strong{anger}: self-reported anger after listening to the 15 minutes of sound from 0 (Maria Taylor) to 100 (Corey Taylor)
}
}
\keyword{datasets}
