% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{metal_health}
\alias{metal_health}
\title{Metal health}
\format{
A tibble with 2506 rows and 2 variables.
}
\source{
\href{https://www.discovr.rocks/csv/metal_health.csv}{www.discovr.rocks/csv/metal_health.csv}
}
\usage{
metal_health
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Lacourse et al. (2001) conducted a study to see whether suicide risk was related to listening to heavy metal music.
They devised a scale to measure preference for bands falling into the category of heavy metal.
This scale included heavy metal bands (Black Sabbath, Iron Maiden), speed metal bands (Slayer, Metallica),
death/black metal bands (Obituary, Burzum) and gothic bands (Marilyn Manson, Sisters of Mercy).
They then used this (and other variables) as predictors of suicide risk based on a scale measuring suicidal ideation etc.
These data are from a fictitious replication. There are two variables representing scores on the scales described above:
\itemize{
\item \strong{hm}: the extent to which the person listens to heavy metal music
\item \strong{suicide}: the extent to which someone has suicidal ideation
}
}
\references{
\itemize{
\item Lacourse, E., Claes, M., & Villeneuve, M. (2001). Heavy metal music and adolescent suicidal risk. \emph{Journal of Youth and Adolescence}, \emph{30}, 321–332. \doi{10.1023/A:1010492128537}
}
}
\keyword{datasets}
