% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{metallica}
\alias{metallica}
\title{Metallica data}
\format{
A tibble with 7 rows and 9 variables.
}
\source{
\href{https://www.discovr.rocks/csv/metallica.csv}{www.discovr.rocks/csv/metallica.csv}
}
\usage{
metallica
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
The data show various pieces of information about past and present members of the band Metallica that may or may not be accurate at the time of writing (2019).
The data contains the following variables:
\itemize{
\item \strong{name}: the band member's name
\item \strong{birth_date}: the band member's date of birth
\item \strong{death_date}: the band member's date of death (where applicable)
\item \strong{instrument}: the instrument played by the band member
\item \strong{current_member}: is the member currently in the band? (True or False)
\item \strong{songs_written}: the number of songs the band member has contributed to
\item \strong{net_worth}: the band member's net worth as of 2019 according to some dodgy website
\item \strong{albums}: the number of studio albums each member played on (up to 2020)
\item \strong{worth_per_song}: the members net worth per song contributed to
}
}
\keyword{datasets}
