% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{methodChoice}
\alias{methodChoice}
\title{splitts calculation in paired and unpaired sections}
\usage{
methodChoice(method, x)
}
\arguments{
\item{method}{is the method parameter and the}

\item{input}{is the input data.frame}
}
\value{
returns the statistical calculations for each section
}
\description{
splitts calculation in paired and unpaired sections
}

