% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{mixed_attitude}
\alias{mixed_attitude}
\title{Imagery and advertising example}
\format{
A tibble with 180 rows and 5 variables
}
\source{
\href{https://www.discovr.rocks/csv/mixed_attitude.csv}{www.discovr.rocks/csv/speed_date.csv}
}
\usage{
mixed_attitude
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A marketing researcher was interested in the effects of types of imagery (positive, negative or neutral)
on perceptions of different types of drink (beer, wine, water). Participants viewed videos of different
drink products in the context of positive, negative or neutral imagery and then rated the products on a
scale from –100 (extremely dislike) through 0 (neutral) to 100 (extremely like).
Those who identify as men and women might respond differently to the products, so participants self-reported
their gender (a between-group variable). The (fictional) data contains the following variables:
\itemize{
\item \strong{id}: participant ID
\item \strong{gender}: gender identity (self-identify as male or female)
\item \strong{drink}: The drink use din the advert (beer, wine or water)
\item \strong{imagery}: The valence of the imagery used in the advert (positive, negative, neutral)
}
}
\keyword{datasets}
