% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{murder}
\alias{murder}
\title{Murder in the streets data}
\format{
A tibble with 36 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/muris_2008.csv}{www.discovr.rocks/csv/murder.csv}
}
\usage{
murder
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictitious data about murder. A sociologist wanted to compare murder rates (murder) each month in a year at
three high-profile locations in London (street). The data contains the following variables:
\itemize{
\item \strong{month}: The month for the reported crime statistics
\item \strong{street}: The street location (Ruskin Avenue, Acacia Avenue and Rue Morgue)
\item \strong{murder}: the number of reported murders during each month
}
}
\keyword{datasets}
