% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{nichols_2004}
\alias{nichols_2004}
\title{Internet addiction scale (IAS) data}
\format{
A tibble with 207 rows and 38 variables.
}
\source{
\href{https://www.discovr.rocks/csv/nichols_2004.csv}{www.discovr.rocks/csv/nichols_2004.csv}
}
\usage{
nichols_2004
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
The increasing populatrity (and usefulness) of the Internet has led to the serious problem of internet addiction.
To research this construct it's helpful to be able to measure it, so Laura Nichols and Richard Nicki developed the
Internet Addiction Scale, IAS (Nichols & Nicki, 2004). This 36-item questionnaire contains items such as \emph{I have stayed on the Internet
longer than I intended to} and \emph{My grades/work have suffered because of my Internet use} to which responses are made on a five-point scale
(\emph{never}, \emph{rarely}, \emph{sometimes}, \emph{frequently}, \emph{always}). The authors dropped two items because they had low means and variances,
and dropped three others because of relatively low correlations with other items. They performed a principal component analysis on the remaining 31 items (\emph{N} = 207).
\itemize{
\item \strong{participant_code}: The participant id
\item \strong{gender}: The participant biological sex
\item \strong{ias1}: responses (1-5) to the question \emph{I find that I need to use the Internet more to get the same enjoyment as before.}
\item \strong{ias2}: responses (1-5) to the question \emph{When I use the Internet now, I do not feel as good as I used to.}
\item \strong{ias3}: responses (1-5) to the question \emph{Time spent on the Internet now is not as enjoyable as it was when I first started using the Internet.}
\item \strong{ias4}: responses (1-5) to the question \emph{Since I first began using the Internet I would say that the amount of time I spend on line has increased but not the satisfaction.}
\item \strong{ias5}: responses (1-5) to the question \emph{I feel depressed, moody or nervous when I am off the internet which goes away when I log on.}
\item \strong{ias6}: responses (1-5) to the question \emph{I feel distressed when I am unable to spend as much time on the Internet as I usually do.}
\item \strong{ias7}: responses (1-5) to the question \emph{The more time I spend away from the Internet, the more irritable I feel.}
\item \strong{ias8}: responses (1-5) to the question \emph{When I attempt to cut back of stop using the Internet I find that the irritability that I experience is relieved by going back on the Internet}
\item \strong{ias9}: responses (1-5) to the question \emph{I have stayed on the Internet  longer than I intended to.}
\item \strong{ias10}: responses (1-5) to the question \emph{I have said to myself 'just a few more minutes on the Internet.'}
\item \strong{ias11}: responses (1-5) to the question \emph{I find myself accessing more information on the Internet that I had planned to.}
\item \strong{ias12}: responses (1-5) to the question \emph{I find myself doing more things on the Internet than I had intended to}
\item \strong{ias13}: responses (1-5) to the question \emph{I have felt a persistent desire to cut down or control my use of the Internet.}
\item \strong{ias14}: responses (1-5) to the question \emph{I have attempted to spend less time on the Internet but I have been unable to do so.}
\item \strong{ias15}: responses (1-5) to the question \emph{I have tried unsuccessfully to restrict my Internet use because of previous over use.}
\item \strong{ias16}: responses (1-5) to the question \emph{I would like to spend less time on the Internet.}
\item \strong{ias17}: responses (1-5) to the question \emph{I have walked or driven to campus/work specifically to use the Internet at times when I normally would not go to campus/work}
\item \strong{ias18}: responses (1-5) to the question \emph{After being on the Internet late into the night in sleep late the next morning because of my Internet use.}
\item \strong{ias19}: responses (1-5) to the question \emph{Once I am on the Internet, I seem to stay on for a long time.}
\item \strong{ias20}: responses (1-5) to the question \emph{I am on the Internet so much that I have to make up for the lost time.}
\item \strong{ias21}: responses (1-5) to the question \emph{I have missed class/work so that I would have more time to spend on the Internet.}
\item \strong{ias22}: responses (1-5) to the question \emph{I have neglected things, which are important and need doing.}
\item \strong{ias23}: responses (1-5) to the question \emph{I see my friends less often because of the time that I spend on the Internet.}
\item \strong{ias24}: responses (1-5) to the question \emph{I have given up a particular recreational activity in order that I would have more time on the Internet}
\item \strong{ias25}: responses (1-5) to the question \emph{At  times I have tried to conceal how long I have been on the Internet}
\item \strong{ias26}: responses (1-5) to the question \emph{My grades/work have suffered because of my Internet use.}
\item \strong{ias27}: responses (1-5) to the question \emph{I have lost sleep because of my Internet use}
\item \strong{ias28}: responses (1-5) to the question \emph{The Internet has affected my life in a negative way.}
\item \strong{ias29}: responses (1-5) to the question \emph{The people I know through the Internet  know me better than my friends at university}
\item \strong{ias30}: responses (1-5) to the question \emph{I prefer socializing on the Internet rather than in person with my friends and family}
\item \strong{ias31}: responses (1-5) to the question \emph{I feel that life without the Internet would be boring and empty.}
\item \strong{ias32}: responses (1-5) to the question \emph{I find myself thinking/longing about when I will go on the Internet again.}
\item \strong{ias33}: responses (1-5) to the question \emph{When I feel lonely, I use the Internet to talk to others.}
\item \strong{ias34}: responses (1-5) to the question \emph{When I use the Internet, I experience a buzz or a high (i.e., feeling elated).}
\item \strong{ias35}: responses (1-5) to the question \emph{I use the Internet as a way of escaping the real world.}
\item \strong{ias36}: responses (1-5) to the question \emph{I use the Internet as a way of escaping the “real world.”}
}
}
\references{
\itemize{
\item Nichols, L. A., & Nicki, R. (2004). Development of a psychometrically sound internet addiction scale: A preliminary step. \emph{Psychology of Addictive Behaviors}, 18, 381–384. \url{https://doi.org/10.1037/0893-164X.18.4.381}
}
}
\keyword{datasets}
