% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{ong_2011}
\alias{ong_2011}
\title{Ong et al. (2011) data: wide/messy format}
\format{
A tibble with 275 rows and 12 variables.
}
\source{
\href{https://www.discovr.rocks/csv/ong_2011.csv}{www.discovr.rocks/csv/ong_2011.csv}
}
\usage{
ong_2011
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A study by Ong et al., (2011) examining the relationship between a person's narcissism and other people's ratings of their profile picture on Facebook.
The pictures were rated on each of four dimensions: coolness, glamour, fashionableness, and attractiveness. In addition, each person was measures on introversion/extroversion and narcissism.
These data are in messy/wide format. The data contains the following variables:
\itemize{
\item \strong{id}: a number identifying he participant
\item \strong{grade}: participants grade at school (Sec 1, Sec 2 or Sec 3)
\item \strong{age}: participant's age in years
\item \strong{sex}: biological sex of the participant
\item \strong{status}: frequency of changing ones Facebook status per week
\item \strong{attractiveness}: rating of profile picture along the dimension of physical attractiveness (1 = not attractive, 5 = very attractive)
\item \strong{fashionable}: rating of profile picture along the dimension of fashionable of profile picture (1 = not fashionable, 5 = very fashionable)
\item \strong{glamour}: rating of profile picture along the dimension of glamour (1 = not glamourous, 5 = very glamourous)
\item \strong{cool}: rating of profile picture along the dimension of cool (1 = not cool, 5 = very cool)
\item \strong{profile}: sum of profile picture ratings
\item \strong{extraversion}: score on the NEO Five-Factor Inventory (NEO-FFI) extraversion scale
\item \strong{narcissism}: score on the Narcissistic Personality Questionnaire for Children-Revised (NPQC-R)
}
}
\references{
\itemize{
\item Ong, E. Y. L., Ang, R. P., Ho, J. C. M., Lim, J. C. Y., Goh, D. H., Lee, C. S., & Chua, A. Y. K. (2011). Narcissism, extraversion and adolescents' self-presentation on Facebook. \emph{Personality and Individual Differences}, 50, 180–185. \doi{10.1016/j.paid.2010.09.022}
}
}
\keyword{datasets}
