% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{pairedTest}
\alias{pairedTest}
\title{calculating paired tests}
\usage{
pairedTest(input)
}
\arguments{
\item{input}{is the data.frame}
}
\value{
returns list with all statistical test results
}
\description{
calculating paired tests
}

