% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticFun.R
\name{pcaReduce}
\alias{pcaReduce}
\title{PCA dimension reduction}
\usage{
pcaReduce(input)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
shows if datadimensionality can be reduced
indicates which percentage of all principal components that have a sigma value > 0.5
}
\description{
PCA dimension reduction
}

