% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{perham_2012}
\alias{perham_2012}
\title{Perham & Sykora (2012) data}
\format{
A tibble with 600 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/perham_2012.csv}{www.discovr.rocks/csv/perham_2012.csv}
}
\usage{
perham_2012
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Perham and Sykora (2012) were interested in the effects of liked and disliked music (compared to quiet) on people's ability to remember things. Twenty-five participants remembered lists of eight letters.
They manipulated the background noise while each list was presented: silence (the control), liked music or disliked music. They used music that they believed most participants would like (a popular song called
'From Paris to Berlin' by Infernal) and dislike (songs such as Repulsion's 'Acid Bath', 'Eaten Alive' and 'Splattered Cadavers' – in other words, the sort of thing I listen to,
although I don't actually have any stuff by Repulsion). Participants recalled each list of eight letters, and the authors calculated the probability of correctly recalling a letter in each position in the list. The data contains the following variables:
\itemize{
\item \strong{id}: Participant's ID (added by me, not from the original data)
\item \strong{sound}: Was the background noise silence (the control), liked music or disliked music
\item \strong{position}: Position in the list of the letter from 1 to 8
\item \strong{recall}: the probability of recalling the letter
}
}
\references{
\itemize{
\item Perham, N., & Sykora, M. (2012). Disliked music can be better for performance than liked music. \emph{Applied Cognitive Psychology}, 26, 550–555. \doi{10.1002/acp.2826}
}
}
\keyword{datasets}
