% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{pets}
\alias{pets}
\title{Pets and life satisfaction data}
\format{
A tibble with 20 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/pets.csv}{www.discovr.rocks/csv/pets.csv}
}
\usage{
pets
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictitious data relating to whether a fish or a cat is a better pet. Life satisfaction and how much they liked animals was measured in samples of fish (\emph{n} = 12) and cat (\emph{n} = 8) owners.
The data contains the following variables:
\itemize{
\item \strong{id}: The participant's id
\item \strong{pet}: whether the person owned a fish or a cat
\item \strong{animal}: how much the person likes animals
\item \strong{life_satisfaction}: the person's life satisfaction score
}
}
\keyword{datasets}
