% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{piff_2012_pedestrian}
\alias{piff_2012_pedestrian}
\title{Data about social class and driver behaviour towards pedestrians}
\format{
A tibble with 152 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/piff_2012_pedestrian.csv}{www.discovr.rocks/csv/piff_2012_pedestrian.csv}
}
\usage{
piff_2012_pedestrian
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
The behaviour of drivers has been used to claim that people of a higher social class are more unpleasant (Piff et al., 2012).
Piff and colleagues classified social class by the type of car on a five-point scale and observed whether the drivers cut
in front of other cars at a busy intersection (see \link{piff_2012_vehicle}) and in a separate study whether drivers cut off pedestrians at a crossing (these data).
These data are reconstructed from Figure 1B of Piff et al. (2012):
\itemize{
\item \strong{id}: Case ID
\item \strong{vehicle}: observers codes of vehicle status (make, age, and appearance) to index drivers social class from 1 to 5
\item \strong{pedestrian_cut}: Whether the driver cut off pedestrians at a crossing
}
}
\references{
\itemize{
\item Piff, P. K., Stancato, D. M., Côté, S., Mendoza-Dentona, R., & Keltner, D. (2012). Higher social class predicts increased unethical behavior. \emph{Proceedings of the National Academy of Sciences}, 109, 4086–4091. \doi{10.1073/pnas.1716910114}
}
}
\keyword{datasets}
