% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotFun}
\alias{plotFun}
\title{Graphical Output}
\usage{
plotFun(input, style)
}
\arguments{
\item{input}{taking the output of all statistical methods}
}
\value{
returns Graphical output
}
\description{
Graphical Output
}

