% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{profile_pic}
\alias{profile_pic}
\title{Profile picture data}
\format{
A tibble with 80 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/profile_pic.csv}{www.discovr.rocks/csv/profile_pic.csv}
}
\usage{
profile_pic
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A researcher was interested in the effect of profile pictures on social media on unsolicited attention.
She took 40 people who had profiles on a social networking website; 17 of them had a relationship status of 'single'
and the remaining 23 had their status as 'in a relationship'. We asked these people to set their profile picture to a photo
of them on their own (alone) and to count how many friend request they got from random strangers over 3 weeks, then to switch
it to a photo of them very obviously as part of a romantic couple and record their friend requests from random
strangers over 3 weeks. The (fictional) data contains the following variables:
\itemize{
\item \strong{id}: Participant id
\item \strong{rel_status}: Whether the participant's relationship status is 'single' or 'in a relationship'
\item \strong{profile_pic}: Whether the participant's profile picture depicts them alone or as part of a couple
\item \strong{requests}: The number of unsolicited friend requests (in 3 weeks) from random strangers who categorise their sexual orientation such that they are interested in people of the gender of the participant
}
}
\keyword{datasets}
