% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{pubs}
\alias{pubs}
\title{Pub data}
\format{
A tibble with 8 rows and 2 variables.
}
\source{
\href{https://www.discovr.rocks}{www.discovr.rocks}
}
\usage{
pubs
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Data illustrating the difference between an outlier and an influencial case. The data came to me via David Hitchin, and he in turn got it from Dr Richard Roberts.
I have no idea whether it's real or fictitious. The tibble contains the following variables:
\itemize{
\item \strong{pubs}: The number of pubs in a particular district of London
\item \strong{mortality}: The mortality rate in that district
}
}
\keyword{datasets}
