% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{puppy_ancova}
\alias{puppy_ancova}
\title{More puppy therapy data}
\format{
A tibble with 30 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/puppy_ancova.csv}{www.discovr.rocks/csv/puppy_ancova.csv}
}
\usage{
puppy_ancova
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
The researchers who conducted the puppy therapy study in \link{puppies} suddenly realized that a
participant's baseline happiness would affect their happiness scores after puppy therapy and wanted to adjust for this revelation. Therefore,
they repeated the study on different participants, but measured happiness at baseline (pre-treatment) as well as post-therapy. The tibble contains the following variables:
\itemize{
\item \strong{id}: Participant id
\item \strong{dose}: Treatment group to which the participant was randomly assigned (No puppies (control), 15 minutes of puppy therapy, 30 minutes of puppy therapy)
\item \strong{pre_happy}: Self-reported happiness from 0 (as unhappy as I can possibly imagine being) to 10 (as happy as I can possibly imagine being) before puppy therapy began.
\item \strong{post_happy}: Self-reported happiness from 0 (as unhappy as I can possibly imagine being) to 10 (as happy as I can possibly imagine being) after puppy therapy concluded.
}
}
\keyword{datasets}
