% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{puppy_love}
\alias{puppy_love}
\title{More puppy therapy data}
\format{
A tibble with 30 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/puppy_love.csv}{www.discovr.rocks/csv/puppy_love.csv}
}
\usage{
puppy_love
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
The researchers who conducted the puppy therapy study in \link{puppies} suddenly realized that a
participant's love of dogs would affect whether puppy therapy would affect happiness. Therefore,
they repeated the study on different participants, but included a self-report measure of love of puppies from 0 (I am a weird person who hates puppies, please be deeply suspicious of me) to 7 (puppies are the best thing ever, one day I might marry one).  The tibble contains the following variables:
\itemize{
\item \strong{id}: Participant id
\item \strong{dose}: Treatment group to which the participant was randomly assigned (No puppies (control), 15 minutes of puppy therapy, 30 minutes of puppy therapy)
\item \strong{happiness}: Self-reported happiness from 0 (as unhappy as I can possibly imagine being) to 10 (as happy as I can possibly imagine being)
\item \strong{puppy_love}: Self-reported love of puppies from 0 (I am a weird person who hates puppies, please be deeply suspicious of me) to 7 (puppies are the best thing ever, one day I might marry one)
}
}
\keyword{datasets}
