% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{pValue}
\alias{pValue}
\title{P-value extractor function}
\usage{
pValue(input)
}
\arguments{
\item{input}{taking the output of all statistical methods}
}
\value{
extacts p value and adds them to data.frame
}
\description{
P-value extractor function
}

