% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{r_exam}
\alias{r_exam}
\title{R exam data data}
\format{
A tibble with 100 rows and 6 variables.
}
\source{
\href{https://www.discovr.rocks/csv/r_exam.csv}{www.discovr.rocks/csv/r_exam.csv}
}
\usage{
r_exam
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictitious data relating to an R exam at two universities. The tibble contains the following variables:
\itemize{
\item \strong{id}: The student's id
\item \strong{exam}: first-year R exam scores as a percentage
\item \strong{computer}: a measure of computer literacy as a percentage
\item \strong{lecture}: percentage of statistics lectures attended
\item \strong{numeracy}: a measure of numerical ability out of 15
\item \strong{uni}: The university attended (Sussex University or Duncetown University)
}
}
\keyword{datasets}
