% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{raq}
\alias{raq}
\title{R Anxiety Questionnaire (RAQ)}
\format{
A tibble with 2,571 rows and 24 variables.
}
\source{
\href{https://www.discovr.rocks/csv/raq.csv}{www.discovr.rocks/csv/raq.csv}
}
\usage{
raq
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictitious data relating to a fictional questionnaire about R anxiety. I can't stress enough how fictional this example is. Like, don't email me for the questionnaire
the whole thing is figment of my mind (and some data simulation). I thought this would be obvious from the
questions, but apparently not. Imagine that I wanted to design a questionnaire to
measure a trait that I termed 'R anxiety'. I devised a questionnaire to measure various aspects of students'
anxiety towards learning R, the RAQ. I generated (in my imagination) questions based on interviews (that never happened in real life)
with anxious and non-anxious students and came up with 23 possible questions to include. Each question was a statement
followed by a five-point Likert scale: \emph{strongly disagree} = 1, \emph{disagree}, \emph{neither agree nor disagree}, \emph{agree} and \emph{strongly agree} (SD, D, N, A and SA respectively).
What's more, I wanted to know whether anxiety about R could be broken down into specific forms of anxiety. In other words,
what latent variables contribute to anxiety about R?

With a little help from a few lecturer friends (this never happened in real life) I collected 2571 completed questionnaires. The data are stored in this object with 2,571 rows and 24 columns.
\itemize{
\item \strong{id}: The student's id
\item \strong{raq_01}: responses (1-5) to the question \emph{Statistics make me cry}
\item \strong{raq_02}: responses (1-5) to the question \emph{My friends will think I'm stupid for not being able to cope with R}
\item \strong{raq_03}: responses (1-5) to the question \emph{Standard deviations excite me}
\item \strong{raq_04}: responses (1-5) to the question \emph{I dream that Pearson is attacking me with correlation coefficients}
\item \strong{raq_05}: responses (1-5) to the question \emph{I don't understand statistics}
\item \strong{raq_06}: responses (1-5) to the question \emph{I have little experience of computers}
\item \strong{raq_07}: responses (1-5) to the question \emph{All computers hate me}
\item \strong{raq_08}: responses (1-5) to the question \emph{I have never been good at mathematics}
\item \strong{raq_09}: responses (1-5) to the question \emph{My friends are better at statistics than me}
\item \strong{raq_10}: responses (1-5) to the question \emph{Computers are useful only for playing games}
\item \strong{raq_11}: responses (1-5) to the question \emph{I did badly at mathematics at school}
\item \strong{raq_12}: responses (1-5) to the question \emph{People try to tell you that R makes statistics easier to understand but it doesn't}
\item \strong{raq_13}: responses (1-5) to the question \emph{I worry that I will cause irreparable damage because of my incompetence with computers}
\item \strong{raq_14}: responses (1-5) to the question \emph{Computers have minds of their own and deliberately go wrong whenever I use them}
\item \strong{raq_15}: responses (1-5) to the question \emph{Computers are out to get me}
\item \strong{raq_16}: responses (1-5) to the question \emph{I weep openly at the mention of central tendency}
\item \strong{raq_17}: responses (1-5) to the question \emph{I slip into a coma whenever I see an equation}
\item \strong{raq_18}: responses (1-5) to the question \emph{R always crashes when I try to use it}
\item \strong{raq_19}: responses (1-5) to the question \emph{Everybody looks at me when I use R}
\item \strong{raq_20}: responses (1-5) to the question \emph{I can't sleep for thoughts of eigenvectors}
\item \strong{raq_21}: responses (1-5) to the question \emph{I wake up under my duvet thinking that I am trapped under a normal distribution}
\item \strong{raq_22}: responses (1-5) to the question \emph{My friends are better at R than I am}
\item \strong{raq_23}: responses (1-5) to the question \emph{If I am good at statistics people will think I am a nerd}
}
}
\keyword{datasets}
