% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{reality_tv}
\alias{reality_tv}
\title{Reality TV example}
\format{
A tibble with 32 rows and 4 variables
}
\source{
\href{https://www.discovr.rocks/csv/reality_tv.csv}{www.discovr.rocks/csv/speed_date.csv}
}
\usage{
reality_tv
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A researcher hypothesized that reality TV show contestants start off with personality disorders that
are exacerbated by being forced to spend time with people as attention-seeking as them.
To test this hypothesis, she gave eight contestants a questionnaire measuring personality disorders
before and after they entered the show. A second group of eight people were given the questionnaires
at the same time; these people were short-listed to go on the show, but never did. The (fictional) data
contains the following variables:
\itemize{
\item \strong{id}: participant ID
\item \strong{contestant}: whether the participant was a contestant or was on the short list but never went on the show
\item \strong{time}: the time at which personality disorder traits were measured (before or after the show)
\item \strong{pd_score}: the score on a personality disorder traits questionnaire
}
}
\keyword{datasets}
