% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{roaming_cats}
\alias{roaming_cats}
\title{Roaming cats data}
\format{
A tibble with 60 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/roaming_cats.csv}{www.discovr.rocks/csv/roaming_cats.csv}
}
\usage{
roaming_cats
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about roaming cats. I was interested in the relationship between the sex of a cat and how much time it spent away from home.
I had heard that male cats disappeared for substantial amounts of time on long-distance roams around the neighbourhood (something about hormones driving them to find mates) whereas female cats tended to be more homebound.
The data set has four variables
\itemize{
\item \strong{id}: Cat id
\item \strong{time}: Time spent away from home per week
\item \strong{sex}: biological sex of the cat as a factor
}
}
\keyword{datasets}
