% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{rollercoaster}
\alias{rollercoaster}
\title{Roaming cats data}
\format{
A tibble with 20 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/rollercoaster.csv}{www.discovr.rocks/csv/rollercoaster.csv}
}
\usage{
rollercoaster
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data based on a study by Meston & Frohlich (2003) that showed that heterosexual people rate a
picture of someone of the opposite sex as more attractive after riding a roller-coaster compared to before.
Imagine we took 20 people as they came off the Rockit roller-coaster at Universal studios in Orlando
and asked them to rate the attractiveness of people in a series of photographs on a scale of 0
(looks like Jabba the Hut) to 10 (looks like Princess Leia or Han Solo). The mean of their
attractiveness ratings was the outcome. We also recorded their fear during the ride using a device
that collates various indicators of  physiological arousal and returns a value from 0, chill, to 10, terrified.
This variable is the predictor. The prediction was that fear would be positively associated with ratings of attractiveness.
\itemize{
\item \strong{id}: Participant id
\item \strong{attractiveness}: Mean attractiveness rating people in a series of photographs from 0 (Jabba the Hut) to 10 (Princess Leia or Han Solo)
\item \strong{fear}: fear during the ride measured on a device that collates various indicators of physiological arousal into a value from 0, chill, to 10, terrified)
}
}
\keyword{datasets}
