% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{sage_editors_cant_play_football}
\alias{sage_editors_cant_play_football}
\title{Data about SAGE editors and their football prowess}
\format{
A tibble with 4 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/sage_editors_cant_play_football.csv}{www.discovr.rocks/csv/sage_editors_cant_play_football.csv}
}
\usage{
sage_editors_cant_play_football
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about SAGE editors and their football prowess. Certain former editors of mine at SAGE
like to think they're great at football (soccer). To see whether they are better than Sussex lecturers and
postgraduates we invited employees of SAGE to join in our football matches. Every person played in one match.
Over many matches, we counted the number of players that scored goals. The data contains the following variables:
\itemize{
\item \strong{employer}: Whether someone was employed by University of Sussex or SAGE publications
\item \strong{score}: Whether someone scored or not in the game they played
\item \strong{frequency}: the number of people falling into each combination of employer and scored
}
}
\keyword{datasets}
