% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{santas_log}
\alias{santas_log}
\title{Self-help book data}
\format{
A tibble with 400 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/santas_log.csv}{www.discovr.rocks/csv/santas_log.csv}
}
\usage{
santas_log
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Let's begin with a Christmas tale. A year ago Santa was resting in his workshop studying his nice
and naughty lists. He noticed a name on the naughty list in bold, upper case letters.
It said \strong{ANDY FIELD OF UNIVERSITY OF SUSSEX}. He went to look up the file of this Andy Field
character. He stared into his snow globe, and as the mists cleared he saw a sad, lonely,
friend-less character walking across campus. Under one arm a box of chocolates, under
the other a small pink Hippo. As he walked the campus he enticed the young students
around him to follow him by offering chocolate. Like the Pied Piper, he led them to
a large hall. Once inside, the boys and girls' eyes glistened in anticipation of
more chocolate. Instead he unleashed a monologue about the general linear model of
such fearsome tedium that Santa began to wonder how anyone could have grown to be
so soulless and cruel.

Santa dusted off his sleigh and whizzed through the night sky to the Sussex campus.
Once there he confronted the evil fiend that he had seen in his globe. "You've been
a naughty boy," he said. "I give you a choice. Give up teaching statistics, or
I will be forced to let the \href{https://en.wikipedia.org/wiki/Krampus}{Krampus} pay
you a visit."

Andy looked sad, "But I love statistics," he said to Santa, "It's cool."

Santa pulled out a candy cane, from it emerged a screen. Just as he was about to instruct
the screen to call the Krampus, an incoming message appeared: some presents had not been delivered last Christmas!

What was Santa to do? How could he find out what determines whether presents get delivered or not? He panicked.

Just then, Santa heard a sad little voice. It said, "I can help you".

"How? replied Santa.

"My students," he replied, "they can save Christmas. All they need are some data."

With that, Santa looked into his candy screen at the elves who had called him, and turned to Andy. "Tell them what you need."

Andy discovered that to deliver presents Santa uses a large team of elves, and that at each house
they usually consume treats. The treats might be Christmas pudding, or sometimes
mulled wine. He also discovered that they consume different quantities.
Sometimes nothing is left, but other times there might be 1, 2, 3 or even 4
pieces of pudding or glasses of mulled wine. The Elves transmitted a log of
400 of the previous year's deliveries. The (fictional) data contains the following variables:
\itemize{
\item \strong{id}: Name of the elf doing the delivery
\item \strong{quantity}: How many treats the elf ate before attempting the delivery
\item \strong{treat}: which kind of treats were consumed (Christmas pudding or mulled wine)
\item \strong{delivered}: were the presents delivered (delivered or not delivered) The (fictional) data contains the following variables:
}
}
\keyword{datasets}
