% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{schutzwohl_2008}
\alias{schutzwohl_2008}
\title{Perham & Sykora (2012) data}
\format{
A tibble with 480 rows and 8 variables.
}
\source{
\href{https://www.discovr.rocks/csv/schutzwohl_2008.csv}{www.discovr.rocks/csv/schutzwohl_2008.csv}
}
\usage{
schutzwohl_2008
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
People can be jealous when they think that their partner is being unfaithful. An evolutionary view suggests that men and women have evolved distinctive types of jealousy:
specifically, a woman's sexual infidelity deprives her mate of a reproductive opportunity and could burden him with years investing in a child that is not his. Conversely, a man's sexual infidelity does not burden his mate
with unrelated children, but may divert his resources from his mate's progeny. This diversion of resources is signalled by emotional attachment to another female.
Consequently, men's jealousy mechanism should have evolved to prevent a mate's sexual infidelity, whereas in women it has evolved to prevent emotional infidelity.
If this is the case, women should be 'on the lookout' for emotional infidelity, whereas men should be watching out for sexual infidelity.

Whether or not you buy into this theory, it can be tested. Achim Schützwohl exposed men and women to sentences on a computer screen (Schützwohl, 2008). At each trial, participants saw a target sentence that was emotionally neutral
(e.g., 'The gas station is at the other side of the street'). However, before each of these targets, a distractor sentence was presented that could also be affectively neutral, or could indicate sexual infidelity
(e.g., 'Your partner suddenly has difficulty becoming sexually aroused when he and you want to have sex') or emotional infidelity (e.g., 'Your partner doesn't say "I love you" to you any more').
Schützwohl reasoned that if these distractor sentences grabbed a person's attention then (1) they would remember them, and (2) they would not remember the target sentence that came afterwards (because their attentional resources were focused on the distractor).
These effects should show up only in people currently in a relationship. The outcome was the number of sentences that a participant could remember (out of six), and the predictors were whether the person had a partner or not (relationship) whether the trial used a neutral distractor,
an emotional infidelity distractor or a sexual infidelity distractor, and whether the sentence was a distractor or the target following a distractor. The data contains the following variables:
\itemize{
\item \strong{id}: Participant's ID (added by me, not from the original data)
\item \strong{relationship}: Whether the person was with a partner or not
\item \strong{age}: Participant's age
\item \strong{distractor_colour}: was the distractor presented in white or colour?
\item \strong{gender}: the self-reported gender of the participant
\item \strong{distractor_type}: Was the sentence a distractor or a target?
\item \strong{sentence_type}: Was the distractor sentence affectively neutral, emotionally valenced or sexually valenced?
\item \strong{recall}: the number of sentences that a participant could remember (out of six)
}
}
\references{
\itemize{
\item Schützwohl, A. (2008). The disengagement of attentive resources from task-irrelevant cues to sexual and emotional infidelity. \emph{Personality and Individual Differences}, 44(3), 633–644. \doi{10.1016/j.paid.2007.09.022}
}
}
\keyword{datasets}
