% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{self_help}
\alias{self_help}
\title{Self-help book data}
\format{
A tibble with 20 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks}{www.discovr.rocks}
}
\usage{
self_help
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
‘Pop psychology’ books sometimes spout nonsense that is unsubstantiated by science. I took 20 people in
relationships and randomly assigned them to one of two groups. One group read the famous popular psychology
book \emph{Women are from Bras and men are from Penis}, and the other read \emph{Marie Claire}. The outcome variable was
their relationship happiness after their assigned reading. The (fictional) data contains the following variables:
\itemize{
\item \strong{id}: the participant's id
\item \strong{book}: whether the participants read \emph{Women are from bras and men are from penis} or \emph{Marie Claire}
\item \strong{happy}: the participant's relationship happiness after reading the book assigned to them
}
}
\keyword{datasets}
