% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{self_help_dsur}
\alias{self_help_dsur}
\title{Self-help book vs statistics book data}
\format{
A tibble with 1000 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/self_help_dsur.csv}{www.discovr.rocks/csv/self_help_dsur.csv}
}
\usage{
self_help_dsur
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Twaddle and Sons, the publishers of Women are from Bras and men are from Penis, were upset about my claims
that their book was as useful as a paper umbrella. They ran their own experiment (N = 500) in which relationship
happiness was measured after participants had read their book and after reading the book you are currently reading.
(Participants read the books in counterbalanced order with a six-month delay.)  The (fictional) data contains the following variables:
\itemize{
\item \strong{id}: the participant's id
\item \strong{book}: whether relationship happiness was measured after reading \emph{Women are from bras and men are from penis} or after reading \emph{Discovering statistics using R}
\item \strong{happy}: the participant's relationship happiness after reading each book
}
}
\keyword{datasets}
