% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticFun.R
\name{shapiroT}
\alias{shapiroT}
\title{Shapiro Wilks Test}
\usage{
shapiroT(x)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
indicates normal distribution
}
\description{
Shapiro Wilks Test
}

