% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{sharman_2015}
\alias{sharman_2015}
\title{Sharman & Dingle (2015) data}
\format{
A tibble with 117 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/sharman_2015.csv}{www.discovr.rocks/csv/sharman_2015.csv}
}
\usage{
sharman_2015
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
There's a perception that listening to extreme music causes anger and associated behavioural problems.
As an avid Metal fan and fairly non-angry type of person this sterotype bothers me. Luckily science has come to the rescue.
Sharman & Dingle (2015) tested 39 fans of extreme music (metal). Their heart rate was measured at baseline, during a subsequent anger induction
and while subsequently listening to music of their choice (which included a lot of bands listed at various point in the acknowledgements of my books).
They collected subjective measures too, but this data file contains only the heart rate data from the study.
\itemize{
\item \strong{id}: The participant id (the original data had numeric IDs, which I have replaced with randomly generated alpha-numeric codes)
\item \strong{music}: Whether the participant was in the music or silence condition
\item \strong{phase}: Phase of the experiment (baseline, anger-induction, listening to music)
\item \strong{hr}: Heart rate (BPM)
}
}
\references{
\itemize{
\item Sharman, L., & Dingle, G. A. (2015). Extreme metal music and anger processing. Frontiers in Human Neuroscience, 9. \doi{10.3389/fnhum.2015.00272}
}
}
\keyword{datasets}
