% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{shopping}
\alias{shopping}
\title{Shopping and exercise data}
\format{A tibble with 10 rows and 3 variables.}
\source{
\url{https://www.discoveringstatistics.com/books/discovering-statistics-using-r/}
}
\usage{
shopping
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
According to some highly unscientific research done by a UK department store chain and reported in Marie Claire magazine (\url{http://ow.ly/9Dxvy}), shopping is good for you.
They found that the average woman spends 150 minutes and walks 2.6 miles when she shops, burning off around 385 calories. In contrast, men spend only about 50 minutes shopping, covering 1.5 miles.
This was based on strapping a pedometer on a mere 10 participants. Although I don’t have the actual data, some simulated data based on these means are in this file.

\itemize{
\item \strong{sex}: biological sex of the individual
\item \strong{distance}: the distance travelled in miles
\item \strong{time}: the time spent shopping in minutes
\item \strong{sex}: Biological sex of the individual
\item \strong{distance}: The distance travelled in miles
\item \strong{time}: The time spent shopping in minutes
}
}
\keyword{datasets}
