% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{sing_or_guitar}
\alias{sing_or_guitar}
\title{Singers, guitarists and personality data}
\format{
A tibble with 200 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/sing_or_guitar.csv}{www.discovr.rocks/csv/sing_or_guitar.csv}
}
\usage{
sing_or_guitar
}
\description{
A fabricated dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A musicologist was interested in whether the instrument played in a rock band could be predicted by their personality traits.
She measured personality traits in 200 musicians and noted whether they were the singer of guitarist. See also \link{band_personality}.
The data contains the following variables:
\itemize{
\item \strong{id}: Participant's ID
\item \strong{extroversion}: participant's extroversion score (high score = high extroversion)
\item \strong{agreeableness}: participant's agreeableness score (high score = high agreeableness)
\item \strong{instrument}: the instrument played (singer, guitar, bass or drums)
}
}
\keyword{datasets}
