% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{social_media}
\alias{social_media}
\title{Social media and grammar data}
\format{
A tibble with 100 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/social_media.csv}{www.discovr.rocks/csv/social_media.csv}
}
\usage{
social_media
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Imagine we conducted an experiment in which a group of 25 people were encouraged to message their friends and
post on social media using their mobiles over a six-month period. A second group of 25 people were banned from
messaging and social media for the same period by being given armbands that administered painful shocks in the
presence of microwaves (like those emitted from phones).  The outcome was a percentage score on a grammatical test
that was administered both before and after the intervention. The first independent variable was, therefore, social
media use (encouraged or banned) and the second was the time at which grammatical ability was assessed (baseline or
after 6 months). These data are fictional. The object contains the following variables:
\itemize{
\item \strong{id}: participant id
\item \strong{media_use}: Whether the participant was encouraged to use social media or banned from using it
\item \strong{time}: the time at which the grammar test was taken: before social media use was manipulated (baseline) and 6 months later
\item \strong{grammar}: the score on a grammar test (as a percentage)
}
}
\keyword{datasets}
