% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{sonnentag_2012}
\alias{sonnentag_2012}
\title{Data exploring time pressure and work detachment}
\format{
A tibble with 10 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/sonnentag_2012.csv}{www.discovr.rocks/csv/sonnentag_2012.csv}
}
\usage{
sonnentag_2012
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Simulated data to approximate Figure 1 in Sonnentag (2012). Research suggests that people who can switch off from work (detachment) during off-hours are more satisfied with life and have fewer symptoms of psychological strain (Sonnentag, 2012).
Factors at work, such as time pressure, affect your ability to detach when away from work. A study of 1709 employees measured their time pressure at work (no time pressure, low, medium, high and very high time pressure).
The data contains the following variables:
\itemize{
\item \strong{time_pressure}: time pressure at work (no time pressure, low, medium, high and very high time pressure)
\item \strong{detachment}: Degree to which someone can switch off from work during off-hours (low detachment or very low detachment)
\item \strong{frequency}: the number of people falling into each combination of time pressure and detachment
}
}
\references{
\itemize{
\item Sonnentag, S. (2012). Psychological detachment from work during leisure time: The benefits of mentally disengaging from work. \emph{Current Directions in Psychological Science}, 21, 114–118. \doi{10.1177/0963721411434979}
}
}
\keyword{datasets}
