% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{stalker}
\alias{stalker}
\title{Stalking therapy}
\format{
A tibble with 50 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/stalker.csv}{www.discovr.rocks/csv/stalker.csv}
}
\usage{
stalker
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Some fictional data about therapy for stalking. A few years back I was stalked. You'd think they could have found
someone a bit more interesting to stalk, but apparently times were hard. It could have been a lot worse,
but it wasn't particularly pleasant. I imagined a world in which a psychologist tried two different therapies on
different groups of stalkers (25 stalkers in each treatment). To the first group he gave cruel-to-be-kind therapy
(every time the stalkers followed him around, or sent him a letter, the psychologist attacked them with a cattle prod).
The second therapy was psychodyshamic therapy, in which stalkers were hypnotized and regressed into their childhood
to discuss their penis (or lack of penis), their father's penis, their dog's penis, the seventh penis of a seventh penis,
and any other penis that sprang to mind. The psychologist measured the number of hours stalking in one week both before
(stalk_pre) and after (stalk_post) treatment.The object contains the following variables:
\itemize{
\item \strong{id}: Participant's id code
\item \strong{therapy}: Whether the person was assigned to \emph{Cruel to be kind therapy} or \emph{Psychodyshamic} therapy
\item \strong{stalk_pre}: number of hours the person spent stalking in one week before therapy
\item \strong{stalk_post}: number of hours the person spent stalking in one week after therapy
}
}
\keyword{datasets}
