% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{students}
\alias{students}
\title{Students and lecturers data}
\format{
A tibble with 10 rows and 7 variables.
}
\source{
\href{https://www.discovr.rocks/csv/students.csv}{www.discovr.rocks/csv/students.csv}
}
\usage{
students
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Some fictional data about students and lecturers.
The object contains the following variables:
\itemize{
\item \strong{name}: Name of person
\item \strong{birth_date}: Date of birth (Year-month-day)
\item \strong{group}: whether the person is a student or lecturer
\item \strong{friends}: how many friends the person has. That's actual friends, not social media friends.
\item \strong{alcohol}: Units of alcohol consumed per week
\item \strong{income}: income (per anum)
\item \strong{neurotic}: Score on a neuroticism scale
}
}
\keyword{datasets}
