% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticFun.R
\name{studentt}
\alias{studentt}
\title{student t test}
\usage{
studentt(inp, ...)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
indicates whether parameters are from same population - with equal variances.
}
\description{
student t test
}

