% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{superhero}
\alias{superhero}
\title{Superhero data}
\format{
A tibble with 30 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/superhero.csv}{www.discovr.rocks/csv/superhero.csv}
}
\usage{
superhero
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Children wearing superhero costumes are more likely to harm themselves because of the unrealistic impression of
invincibility that these costumes could create. For example, children have reported to hospital with severe injuries
because of trying 'to initiate flight without having planned for landing strategies' (Davies, Surridge, Hole, & Munro-Davies, 2007).
I can relate to the imagined power that a costume bestows upon you; indeed, I have been known to dress up as Fisher by donning a
beard and glasses and trailing a goat around on a lead in the hope that it might make me more knowledgeable about statistics.
These fictional data contain the severity of injury (on a scale from 0, no injury, to 100, death) for children reporting to the accident
and emergency department at hospitals, and information on which superhero costume they were wearing (hero):
Spiderman, Superman, the Hulk or a teenage mutant ninja turtle. The fictitious data contain the following variables:
\itemize{
\item \strong{id}: The participant's id
\item \strong{hero}: The costume being worn at the time of injury (Spiderman, Superman, the Hulk or a teenage mutant ninja turtle)
\item \strong{injury}: the severity of injury (on a scale from 0, no injury, to 100, death)
}
}
\keyword{datasets}
