% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{supermodel}
\alias{supermodel}
\title{Supermodel data}
\format{
A tibble with 231 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/supermodel.csv}{www.discovr.rocks/csv/supermodel.csv}
}
\usage{
supermodel
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A fashion student was interested in factors that predicted the salaries of male and female catwalk models.
She collected data from 231 models (supermodel.csv). For each model she asked them their salary per day (salary),
their age (age), their length of experience as models (years), and their industry status as a model as their
percentile position rated by a panel of experts (status). The fictitious data contain the following variables:
\itemize{
\item \strong{salary}: The model's salary
\item \strong{age}: The model's age (years)
\item \strong{years}: The model's experience (years in the industry)
\item \strong{status}: Model's status as their percentile position (\%) rated by a panel of experts.
}
}
\keyword{datasets}
