% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{switch}
\alias{switch}
\title{Switch: games console injuries}
\format{
A tibble with 120 rows and 5 variables.
}
\source{
\href{https://www.discovr.rocks/csv/switch.csv}{www.discovr.rocks/csv/switch.csv}
}
\usage{
switch
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about injuries while playing video games on a console. There are reports of increases in injuries related to
playing games consoles. These injuries were attributed mainly to muscle and tendon strains. A researcher hypothesized
that a stretching warm-up before playing games would help lower injuries, and that athletes would be less susceptible to
injuries because their regular activity makes them more flexible. She took 60 athletes and 60 non-athletes (athlete);
half of them played on a Nintendo Switch and half watched others playing as a control (switch), and within these groups half did a
5-minute stretch routine before playing/watching whereas the other half did not (stretch).
The outcome was a pain score out of 10 (where 0 is no pain, and 10 is severe pain) after playing for 4 hours (injury).
\itemize{
\item \strong{id}: Participant's id
\item \strong{athlete}: Whether the participant was an athlete or not
\item \strong{stretch}: Whether the participant warmed up with stretching (or not)
\item \strong{switch}: Whether the participant played Nintendo Switch games or watched someone else playing
\item \strong{injury}: Injury severity (where 0 is no pain, and 10 is severe pain)
}
}
\keyword{datasets}
