% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{t.test.matrix}
\alias{t.test.matrix}
\title{t-test matrix function}
\usage{
\method{t}{test.matrix}(mat, ...)
}
\arguments{
\item{x}{A data.frame or data.table}
}
\value{
returns the p value matrix of t.tests
this code function was found at http://www.sthda.com/english/wiki/matrix-of-student-t-test
}

