% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{tablets}
\alias{tablets}
\title{Tablet sales data}
\format{
A tibble with 240 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/tablets.csv}{www.discovr.rocks/csv/tablets.csv}
}
\usage{
tablets
}
\description{
A dataset from Field, A. P. (2021). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
A company owner was interested in how to make his brand of (computer) tablets more desirable.
He collected data on how cool people perceived a product's advertising to be, how cool they thought the product was,
and how desirable they found the product. Am I showing my age by using the word 'cool'? The fictitious data contain the following variables:
\itemize{
\item \strong{id}: Participant ID
\item \strong{advert_cool}: Perceived 'coolness' of the advertising campaign from 0 (as cool as Andy Field) to 5 (as cool as something that makes you go 'wow, that's sick', or whatever it is that people under the age of 25 say these days)
\item \strong{desirability}: The desirability of the product from (0 as desirable as Andy Field) to 10 (I \emph{really} want one of those)
\item \strong{product_cool}: Perceived 'coolness' of the product from from 0 (designed by Andy Field) to 5 (Designed by Apple).
}
}
\keyword{datasets}
