% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{tea15}
\alias{tea15}
\title{Tea data (small sample)}
\format{A tibble with 15 rows and 2 variables:}
\source{
\url{https://www.discoveringstatistics.com/books/discovering-statistics-using-r/}
}
\usage{
tea15
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
One of my favourite activities, especially when trying to do brain-melting things like writing statistics books, is drinking tea. I am English, after all.
Fortunately, tea improves your cognitive function – well, it does in old Chinese people at any rate (Feng, Gwee, Kua, & Ng, 2010).
I may not be Chinese and I’m not that old, but I nevertheless enjoy the idea that tea might help me think. Here are some (fictional) data based on Feng et al.’s
study that measured the number of cups of tea drunk per day and cognitive functioning (out of 80) in 15 people.

\itemize{
\item \strong{tea}: the number of cups of tea a person drinks per day
\item \strong{cog_fun}: cognitive functioning (out of 80)
\item \strong{tea}: The number of cups of tea a person drinks per day
\item \strong{cog_fun}: Cognitive functioning (out of 80)
}
}
\references{
\itemize{
\item Feng, L., Gwee, X., Kua, E. H., & Ng, T. P. (2010). Cognitive function and tea consumption in community dwelling older Chinese in Singapore. \emph{Journal of Nutrition Health & Aging}, \emph{14}, 433-438.
}
}
\keyword{datasets}
