% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{teach_method}
\alias{teach_method}
\title{Method of teaching data (3 groups)}
\format{
A tibble with 30 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/teach_method.csv}{www.discovr.rocks/csv/teach_method.csv}
}
\usage{
teach_method
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
To test how different teaching methods affected students' knowledge I took three statistics modules where
I taught the same material. For one module I wandered around with a large cane and beat anyone who asked
daft questions or got questions wrong (punish). In the second I encouraged students to discuss things that
they found difficult and gave anyone working hard a nice sweet (reward). In the final course I neither
punished nor rewarded students' efforts (indifferent). I measured the students' exam marks (percentage).
This fictional data contains the following variables
\itemize{
\item \strong{id}: participant's id
\item \strong{group}: The type of teaching method used (Punish, Reward, Indifferent)
\item \strong{exam}: The exam mark (\%)
}
}
\keyword{datasets}
