% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{teaching}
\alias{teaching}
\title{Method of teaching data}
\format{A tibble with 20 rows and 3 variables:
\describe{
  \item{method}{The type of teaching method used}
  \item{Sex}{Biological sex of the individual}
  \item{mark}{The score out of 20 on a test}
}}
\source{
\url{https://www.discoveringstatistics.com/books/discovering-statistics-using-r/}
}
\usage{
teaching
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
The data show the score (out of 20) for 20 different students, some of whom are male and some female,
and some of whom were taught using positive reinforcement (being nice) and others who were taught using punishment (electric shock)
}
\keyword{datasets}
