% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{text_messages}
\alias{text_messages}
\title{Messaging apps and grammar example}
\format{
A tibble with 100 rows and 4 variables
}
\source{
\href{https://www.discovr.rocks/csv/text_messages.csv}{www.discovr.rocks/csv/speed_date.csv}
}
\usage{
text_messages
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Text messaging and Twitter encourage communication using abbreviated forms of words (if u no wat I mean).
A researcher wanted to see the effect this had on children’s understanding of grammar.
One group of 25 children was encouraged to send text messages on their mobile phones over a 6-month period.
A second group of 25 was forbidden from sending text messages for the same period (to ensure adherence,
this group were given armbands that administered painful shocks in the presence of a phone signal).
The outcome was a score on a grammatical test (as a percentage) that was measured both before and after the experiment.
The (fictional) data contains the following variables:
\itemize{
\item \strong{id}: participant ID
\item \strong{group}: whether the participant was assigned to the text message group or control group
\item \strong{time}: the time at which grammar ability was measured (baseline or 6 months later)
\item \strong{grammar}: the score on the grammar test as a percentage (\%)
}
}
\keyword{datasets}
