% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{tosser}
\alias{tosser}
\title{The Teaching of Statistics for Scientific Experiments—Revised (TOSSE-R) data}
\format{
A tibble with 239 rows and 29 variables.
}
\source{
\href{https://www.discovr.rocks/csv/tosser.csv}{www.discovr.rocks/csv/tosser.csv}
}
\usage{
tosser
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictitious data relating to a fictional questionnaire about The Teaching of Statistics for Scientific Experiments. Again, I stress that this example is fictional.
I thought the name of the questionnaire would give it away, I mean, no-one is calling a questionnaire TOSSER are they? Don't email me for the questionnaire, it's all made up, you definitley
don't want to base your research upon it. Imagine I wanted to revise the 'Teaching of Statistics for Scientific Experiments' (TOSSE) questionnaire, which is (I mean, it isn't because I made it up)
based on Bland's theory that says that good research methods lecturers should have: (1) a profound love of statistics; (2) an enthusiasm for experimental design; (3) a love of teaching; and (4)
a complete absence of normal interpersonal skills. These characteristics should be related (i.e., correlated).
The revised version of this questionnaire (TOSSE – R) was given to 239 research methods lecturers to see if it supported Bland's theory. Each question was a statement
followed by a five-point Likert scale: \emph{strongly disagree} = 1, \emph{disagree}, \emph{neither agree nor disagree}, \emph{agree} and \emph{strongly agree} (SD, D, N, A and SA respectively).
The data contains the following variables
\itemize{
\item \strong{id}: The student's id
\item \strong{q_01}: responses (1-5) to the question \emph{I once woke up in the middle of a vegetable patch hugging a turnip that I'd mistakenly dug up thinking it was Roy's largest root}
\item \strong{q_02}: responses (1-5) to the question \emph{Students are like irritating pigeons pecking away at my sanity}
\item \strong{q_03}: responses (1-5) to the question \emph{I memorize probability values for the F-distribution}
\item \strong{q_04}: responses (1-5) to the question \emph{I worship at the shrine of Pearson}
\item \strong{q_05}: responses (1-5) to the question \emph{I still live with my mother and have little personal hygiene}
\item \strong{q_06}: responses (1-5) to the question \emph{Teaching others makes me want to swallow a large bottle of bleach because the pain of my burning oesophagus would be light relief in comparison}
\item \strong{q_07}: responses (1-5) to the question \emph{Helping others to understand sums of squares is a great feeling}
\item \strong{q_08}: responses (1-5) to the question \emph{I like control conditions}
\item \strong{q_09}: responses (1-5) to the question \emph{I calculate 3 ANOVAs in my head before getting out of bed every morning}
\item \strong{q_10}: responses (1-5) to the question \emph{I could spend all day explaining statistics to people}
\item \strong{q_11}: responses (1-5) to the question \emph{I like it when people tell me I've helped them to understand factor rotation}
\item \strong{q_12}: responses (1-5) to the question \emph{People fall asleep as soon as I open my mouth to speak}
\item \strong{q_13}: responses (1-5) to the question \emph{Designing experiments is fun}
\item \strong{q_14}: responses (1-5) to the question \emph{I'd rather think about appropriate dependent variables than meet people}
\item \strong{q_15}: responses (1-5) to the question \emph{I soil my pants with excitement at the mention of Factor Analysis}
\item \strong{q_16}: responses (1-5) to the question \emph{Thinking about whether to use repeated- or independent-measures thrills me}
\item \strong{q_17}: responses (1-5) to the question \emph{I enjoy sitting in the park contemplating whether to use participant observation in my next experiment}
\item \strong{q_18}: responses (1-5) to the question \emph{Standing in front of 300 people in no way makes me lose control of my bowels}
\item \strong{q_19}: responses (1-5) to the question \emph{I like to help students}
\item \strong{q_20}: responses (1-5) to the question \emph{Passing on knowledge is the greatest gift you can bestow an individual}
\item \strong{q_21}: responses (1-5) to the question \emph{Thinking about Bonferroni corrections gives me a tingly feeling in my groin}
\item \strong{q_22}: responses (1-5) to the question \emph{I quiver with excitement when thinking about designing my next experiment}
\item \strong{q_23}: responses (1-5) to the question \emph{I often spend my spare time talking to the pigeons ... and even they die of boredom}
\item \strong{q_24}: responses (1-5) to the question \emph{I tried to build myself a time machine so that I could go back to the 1930s and follow Fisher around on my hands and knees licking the floor on which he'd just trodden}
\item \strong{q_25}: responses (1-5) to the question \emph{I love teaching}
\item \strong{q_26}: responses (1-5) to the question \emph{I spend lots of time helping students}
\item \strong{q_27}: responses (1-5) to the question \emph{I love teaching because students have to pretend to like me or they'll get bad marks}
\item \strong{q_28}: responses (1-5) to the question \emph{My cat is my only friend}
}
}
\keyword{datasets}
