% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{tumour}
\alias{tumour}
\title{Mobile phone use and brain tumour data}
\format{
A tibble with 102 rows and 3 variables
}
\source{
\url{https://www.discovr.rocks}
}
\usage{
tumour
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Mobile phones emit microwaves, and so holding one next to your brain for large parts of the day is a bit like
sticking your brain in a microwave oven and pushing the ‘cook until well done’ button. If we wanted to test this
experimentally, we could get six groups of people and strap a mobile phone on their heads, then by remote control
turn the phones on for a certain amount of time each day. After six months, we measure the size of any tumour (in mm^3)
close to the site of the phone antenna (just behind the ear). The six groups experienced 0, 1, 2, 3, 4 or 5 hours per
day of phone microwaves for six months. The fictitious data contains three variables:
\itemize{
\item \strong{id}: participant ID
\item \strong{usage}: how many hours per day were the phones active for (0, 1, 2, 3, 4, or 5 hours)
\item \strong{tumour}: Size of any tumour (in mm^3)
}
}
\keyword{datasets}
