% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{tutor_marks}
\alias{tutor_marks}
\title{Tutor marking data}
\format{
A tibble with 32 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/teach_method.csv}{www.discovr.rocks/csv/tutor_marks.csv}
}
\usage{
tutor_marks
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
It is common that lecturers obtain reputations for being ‘hard’ or ‘light’ markers, but there is often little
to substantiate these reputations. A group of students investigated the consistency of marking by submitting the
same essays to four different lecturers. The outcome was the percentage mark given by each lecturer and the
predictor was the lecturer who marked the report. The fictitious data contains three variables:
\itemize{
\item \strong{id}: participant's id
\item \strong{tutor}: The tutor who marked the work
\item \strong{exam}: The mark on the essay (\%)
}
}
\keyword{datasets}
