% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{unPairedTest}
\alias{unPairedTest}
\title{Output for only un-paired tests}
\usage{
unPairedTest(input)
}
\arguments{
\item{input}{taking the output of all statistical methods}
}
\value{
extacts p value of all un-paired tests and adds them to data.frame
}
\description{
Output for only un-paired tests
}

