% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{video_games}
\alias{video_games}
\title{Video game and aggression data}
\format{
A tibble with 442 rows and 4 variables
}
\source{
\url{https://www.discovr.rocks}
}
\usage{
video_games
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Video games are among the favourite online activities for young people. These games have been linked to increased aggression in youths.
Another predictor of aggression and conduct problems is callous-unemotional traits such as lack of guilt, lack of empathy, callous use of others for personal gain.
Imagine that a scientist explored the relationship between playing violent video games and aggression. She measured aggressive behaviour, callous-traits, and the
number of hours per week they play video games in 442 youths. These fictitious data contains three variables:
\itemize{
\item \strong{id}: participant ID
\item \strong{agress}: a measure of aggressive behaviour from 0 (no aggression at all) to 100 (extremely aggressive)
\item \strong{vid_game}: number of hours per week spent playing video games
\item \strong{caunts}: callous unemotional traits measured on the Inventory of Callous-Unemotional Traits (ICU), ranging from 0 (none) to 72 (extreme)
}
}
\keyword{datasets}
