% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{welchTest}
\alias{welchTest}
\title{Welchs two sample T.Test}
\usage{
welchTest(input)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
indicates whether parameter are from same population
}
\description{
Welchs two sample T.Test
}

