% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{williams}
\alias{williams}
\title{Williams' questionnaire of organizational ability data}
\format{
A tibble with 239 rows and 29 variables.
}
\source{
\href{https://www.discovr.rocks/csv/williams.csv}{www.discovr.rocks/csv/williams.csv}
}
\usage{
williams
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Dr Sian Williams (University of Brighton) devised a questionnaire to measure organizational ability.
She predicted five factors to do with organizational ability: (1) preference for organization; (2)
goal achievement; (3) planning approach; (4) acceptance of delays; and (5) preference for routine.
These dimensions are theoretically independent. Williams's questionnaire contains 28 items using a
seven-point Likert scale (1 = \emph{strongly disagree}, 4 = \emph{neither}, 7 = \emph{strongly agree}).
She gave it to 239 people.
\itemize{
\item \strong{participant}: The participant id
\item \strong{sex}: The participant biological sex
\item \strong{org1}: responses (1-7) to the question \emph{I like to have a plan to work to in everyday life}
\item \strong{org2}: responses (1-7) to the question \emph{I feel frustrated when things don't go to plan}
\item \strong{org3}: responses (1-7) to the question \emph{I get most things done in a day that I want to}
\item \strong{org4}: responses (1-7) to the question \emph{I stick to a plan once I have made it}
\item \strong{org6}: responses (1-7) to the question \emph{I enjoy spontaneity and uncertainty}
\item \strong{org7}: responses (1-7) to the question \emph{I feel frustrated if I can't find something I need}
\item \strong{org9}: responses (1-7) to the question \emph{I find it difficult to follow a plan through}
\item \strong{org10}: responses (1-7) to the question \emph{I am an organized person}
\item \strong{org11}: responses (1-7) to the question \emph{I like to know what I have to do in a day}
\item \strong{org12}: responses (1-7) to the question \emph{Disorganized people annoy me}
\item \strong{org13}: responses (1-7) to the question \emph{I leave things to the last minute}
\item \strong{org14}: responses (1-7) to the question \emph{I have many different plans relating to the same goal}
\item \strong{org16}: responses (1-7) to the question \emph{I like to have my documents filed and in order}
\item \strong{org17}: responses (1-7) to the question \emph{I find it easy to work in a disorganized environment}
\item \strong{org18}: responses (1-7) to the question \emph{I make to do lists and achieve most of the things on it}
\item \strong{org19}: responses (1-7) to the question \emph{My workspace is messy and disorganized}
\item \strong{org20}: responses (1-7) to the question \emph{I like to be organized}
\item \strong{org21}: responses (1-7) to the question \emph{Interruptions to my daily routine annoy me}
\item \strong{org22}: responses (1-7) to the question \emph{I feel that I am wasting my time}
\item \strong{org23}: responses (1-7) to the question \emph{I forget the plans I have made}
\item \strong{org24}: responses (1-7) to the question \emph{I prioritize the things I have to do}
\item \strong{org25}: responses (1-7) to the question \emph{I like to work in an organized environment}
\item \strong{org26}: responses (1-7) to the question \emph{I feel relaxed when I don't have a routine}
\item \strong{org27}: responses (1-7) to the question \emph{I set deadlines for myself and achieve them}
\item \strong{org28}: responses (1-7) to the question \emph{I change rather aimlessly from one activity to another during the day}
\item \strong{org29}: responses (1-7) to the question \emph{I have trouble organizing the things I have to do}
\item \strong{org30}: responses (1-7) to the question \emph{I put tasks off to another day}
\item \strong{org31}: responses (1-7) to the question \emph{I feel restricted by schedules and plans}
}
}
\keyword{datasets}
