% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{xbox}
\alias{xbox}
\title{Xbox: games console injuries}
\format{
A tibble with 40 rows and 4 variables.
}
\source{
\href{https://www.discovr.rocks/csv/xbox.csv}{www.discovr.rocks/csv/xbox.csv}
}
\usage{
xbox
}
\description{
A dataset from Field, A. P. (2023). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictional data about injuries while playing video games on a console. A researcher was interested in
what factors contributed to injuries resulting from game console use. She tested 40 participants who
were randomly assigned to either an active or static game played on either a Nintendo Switch or
Xbox One Kinect. At the end of the session their physical condition was evaluated on an
injury severity scale.
\itemize{
\item \strong{id}: Participant's id
\item \strong{game}: Whether the participant played an active or static game
\item \strong{console}: The games console used (Nineto Switch or Xbox Kinect)
\item \strong{injury}: Injury severity (a score ranging from 0 (no injury) to 20 (severe injury))
}
}
\keyword{datasets}
