% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{zibarras_2008}
\alias{zibarras_2008}
\title{Zibarras et al. (2008) data}
\format{
A tibble with 207 rows and 12 variables.
}
\source{
\href{https://www.discovr.rocks/csv/zibarras_2008.csv}{www.discovr.rocks/csv/zibarras_2008.csv}
}
\usage{
zibarras_2008
}
\description{
A dataset from Field, A. P. (2022). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Zibarras, Port, and Woods (2008) looked at the relationship between personality and creativity.
They used the Hogan Development Survey (HDS), which measures 11 dysfunctional dispositions of employed adults:
being \emph{volatile}, \emph{mistrustful}, \emph{cautious}, \emph{detached}, \emph{passive_aggressive}, \emph{arrogant},
\emph{manipulative}, \emph{dramatic}, \emph{eccentric}, \emph{perfectionist}, and \emph{dependent}.
\itemize{
\item \strong{id}: The participant id
\item \strong{volatile}: responses to the question items of the HDS relating to the \emph{volatile} disposition.
\item \strong{mistrustful}: responses to the question items of the HDS relating to the \emph{mistrustful} disposition.
\item \strong{cautious}: responses to the question items of the HDS relating to the \emph{cautious} disposition.
\item \strong{detached}: responses to the question items of the HDS relating to the \emph{detached} disposition.
\item \strong{passive_aggressive}: responses to the question items of the HDS relating to the \emph{passive_aggressive} disposition.
\item \strong{arrogant}: responses to the question items of the HDS relating to the \emph{arrogant} disposition.
\item \strong{manipulative}: responses to the question items of the HDS relating to the \emph{manipulative} disposition.
\item \strong{dramatic}: responses to the question items of the HDS relating to the \emph{dramatic} disposition.
\item \strong{eccentric}: responses to the question items of the HDS relating to the \emph{eccentric} disposition.
\item \strong{perfectist}: responses (1-5) to the question \emph{I have said to myself 'just a few more minutes on the Internet.'}
\item \strong{dependent}: responses (1-5) to the question \emph{I find myself accessing more information on the Internet that I had planned to.}
}
}
\references{
\itemize{
\item Zibarras, L. D., Port, R. L., & Woods, S. A. (2008). Innovation and the 'dark side' of personality: Dysfunctional traits and their relation to self-reported innovative characteristics. \emph{Journal of Creative Behavior}, 42, 201–215. \doi{10.1002/j.2162-6057.2008.tb01295.x}
}
}
\keyword{datasets}
