% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{corTest}
\alias{corTest}
\title{Correlation Test}
\usage{
corTest(input)

corTest(input)
}
\arguments{
\item{input}{A data.frame or data.table}

\item{input}{A data.frame or data.table}
}
\value{
indicates correlation between parameters

analyzes the variance of the given samples
}
\description{
Correlation Test

Anova Test
}

