% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{disc}
\alias{disc}
\title{discovr()}
\usage{
disc(x, method = "unPaired", preset = NULL)
}
\arguments{
\item{x}{A data.frame or data.table}

\item{method}{The method indicates "paired" or "unPaired" testing}

\item{preset}{As preset mutliple techniques are allowed, such as "FACS" or "FRAP".}
}
\value{
The main disc() function returns the htmlwidget graphic
}
\examples{
disc(mtcars)
}
\author{
Robin Kohze
}

