% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{disc}
\alias{disc}
\title{discovr()}
\usage{
disc(x, method = "unPaired", preset = NULL)
}
\arguments{
\item{x}{A data.frame or data.table}

\item{method}{The method indicates "paired" or "unpaired" testing}

\item{preset}{As preset mutliple techniques are allowed, such as "FACS" or "FRAP".}

\item{style}{A style input such as "heatmap" / "cleveland" / "ternary"}
}
\value{
The calculation output
}
\examples{
disc(mtcars[1:2])
}

