% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{disc}
\alias{disc}
\title{discovr()}
\usage{
disc(x, method = "unpaired", preset = NULL, style = "heatmap")
}
\arguments{
\item{x}{A data.frame or data.table}

\item{method}{The method indicates paired or un-paired testing}

\item{preset}{As preset mutliple techniques are allowed, such as FACS or FRAP.}
}
\value{
The calculation output
}
\examples{
library(Discovr)
disc(mtcars[1:2])
}

