% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\docType{package}
\name{discovr}
\alias{discovr}
\title{discovr: Resources for Discovering Statistics Using R and RStudio (Field, 2021)}
\description{
The \code{discovr} package contains interactive \code{learnr} tutorials and datasets that accompany my 2021 textbook \href{https://www.discovr.rocks/}{Discovering Statistics Using R and RStudio}.
}
\section{Who is the package aimed at?}{


Anyone teaching from or reading \href{https://www.discovr.rocks/}{Discovering Statistics Using R and RStudio} should find these resources useful.
}

\section{Interactive tutorials}{


\strong{Getting started}:

I recommend working through \href{http://milton-the-cat.rocks/learnr/r/r_getting_started/}{this tutorial} on how to install, set up and work within R and RStudio before starting the interactive tutorials.

\strong{Running a tutorial}:

To run each tutorial execute\preformatted{learnr::run_tutorial("name_of_tutorial", package = "discovr")
}

Replacing \code{name_of_tutorial} with the name in bold below. For example, to load the tutorial \code{discovr_02} execute:\preformatted{learnr::run_tutorial("discovr_02", package = "discovr")
}
\itemize{
\item \strong{discovr_01}: Key concepts in R (functions and objects, packages and functions, style, data types, tidyverse, tibbles)
\item \strong{discovr_02}: Summarizing data (frequency distributions, grouped frequency distributions, relative frequencies, histograms, mean, median, variance, standard deviation, interquartile range)
\item \strong{discovr_03}: Confidence intervals: interactive app demonstrating what a confidence interval is, computing normal and bootstrap confidence intervals using R, adding confidence intervals to data summaries
\item \strong{discovr_05}: Visualizing data. The ggplot2 package, boxplots, plotting means, violin plots, scatterplots, grouping by colour, grouping using facets, adjusting scales, adjusting positions.
\item \strong{discovr_06}: The beast of bias. Restructuring data from messy to tidy format (and back). Spotting outliers using histograms and boxplots. Calculating z-scores (standardizing scores). Writing your own function. Using z-scores to detect outliers. Q-Q plots. Calculating skewness, kurtosis and the number of valid cases. Grouping summary statistics by multiple categorical/grouping variables.
\item \strong{discovr_07}: Associations. Plotting data with GGally. Pearson's r, Spearman's Rho, Kendall's tau, robust correlations.
\item \strong{discovr_08}: The general linear model (GLM). Visualizing the data, fitting GLMs with one and two predictors. Viewing model parameters with broom, model parameters, standard errors, confidence intervals, fit statistics, significance.
\item \strong{discovr_09}: Categorical predictors with two categories (comparing two means). Comparing two independent means, comparing two related means, effect sizes.
\item \strong{discovr_10}: Moderation and mediation. Centring variables (grand mean centring), specifying interaction terms, moderation analysis, simple slopes analysis, Johnson-Neyman intervals, mediation with one predictor, direct and indirect effects, mediation using lavaan.
\item \strong{discovr_11}: Comparing several means. Essentially 'One-way independent ANOVA' but taught using a general linear model framework. Covers setting contrasts (dummy coding, contrast coding, and linear and quadratic trends), the \emph{F}-statistic and Welch's robust \emph{F}, robust parameter estimation, heteroscedasticity-consistent tests of parameters, robust tests of means based on trimmed data, \emph{post hoc} tests, Bayes factors.
\item \strong{discovr_12}: Comparing means adjusted for other variables. Essentially 'Analysis of Covariance (ANCOVA)' designs but taught using a general linear model framework. Covers setting contrasts, Type III sums of squares, the \emph{F}-statistic, robust parameter estimation, heteroscedasticity-consistent tests of parameters, robust tests of adjusted means, \emph{post hoc} tests, Bayes factors.
\item \strong{discovr_13}: Factorial designs. Fitting models for two-way factorial designs (independent measures) using both lm() and the afex package. This tutorial builds on previous ones to show how models can be fit with two categorical predictors to look at the interaction between them. We look at fitting the models, setting contrasts for the two categorical predictors, obtaining estimated marginal means, interaction plots, simple effects analysis, diagnostic plots, partial eta-squared and partial omega-squared, robust models and Bayes factors.
}

\strong{Workflow}:
The tutorials are self-contained (you practice code in code boxes) so you don't need to use RStudio at the
same time. However, to get the most from them I would recommend that you create an RStudio project and within that open (and save) a new
R Markdown file each time to work through a tutorial. Within that Markdown file, replicate parts of the code from the tutorial (in code chunks)
and use Markdown to write notes about what you have done, and to reflect on things that you have struggled with, or note useful tips to help you
remember things. Basically, write a learning journal. This workflow has the advantage of not just teaching you the code that you need to do certain things,
but also provides practice in using RStudio itself.
}

\section{Datasets}{


See the book or data descriptions for more details. This is a list of available datasets within the package. Raw CSV files are available from the book's website.
\itemize{
\item \link{acdc}: Data about whether Bon Scott or Brian Johnson is the best singer of AC/DC.
\item \link{album_sales}: Fictitious data about predicting album sales from advertising, airplay and the band's image.
\item \link{alien_scents}: Fictitious data about training sniffer dogs to detect alien space lizards when they try to mask their identity with different scents. See also \link{sniffer_dogs}.
\item \link{animal_bride}: Fictitious data about life satisfaction when married to a dog or a goat.
\item \link{beckham_1929}: Data from a study by Beckham (1929).
\item \link{big_hairy_spider}: Fictitious data about whether anxiety is greater after exposure to real spiders or pictures of spiders.
\item \link{biggest_liar}: Fictitious data about creativity and telling lies.
\item \link{bronstein_2019}: Data about whether delusion proneness predicts belief in fake news because of less analytic thinking.
\item \link{bronstein_miss_2019}: The data in \link{bronstein_2019} but with missing values inserted using MCAR amputation.
\item \link{catterplot}: Fictitious data for plotting a catterplot.
\item \link{cat_dance}: Fictitious data about training cats to dance.
\item \link{cat_reg}: Fictitious data about training cats to dance.
\item \link{cetinkaya_2006}: data from a study by Cetinkaya and Domjan (2006) about quails with sexual fetishes. Seriously.
\item \link{chamorro_premuzic}: Data about what students want (personality wise) from their lecturers.
\item \link{child_aggression}: Fictitious data (based on real research) about predicting aggression in children.
\item \link{coldwell_2006}: Data predicting childhood adjustment from various parenting variables.
\item \link{daniels_2012}: Data about the effects of sexualised sports images on self-image.
\item \link{dark_lord}: Fictitious data about the subliminal messages in songs.
\item \link{download}: Fictitious data about the download music festival and being smelly.
\item \link{df_beta}: Fictitious data used to illustrate the DF Beta statistic.
\item \link{eel}: Fictitious data about a randomized control trial to test whether eel therapy is an effective treatment of constipation.
\item \link{elephooty}: Fictitious data about elephants playing football (soccer).
\item \link{essay_marks}: Fictitious data about essay marking.
\item \link{exam_anxiety}: Fictitious data about exam performance, anxiety and revision.
\item \link{gallup_2003}: Data that tests a hypothesis about why penises have a bell end.
\item \link{gelman_2009}: Data used to critically evaluate the explanations (and claim) that there are more beautiful women than men in the world.
\item \link{glastonbury}: More fictitious data about music festivals and being smelly.
\item \link{goggles}: Fictitious data about whether alcohol affects perception of physical attractiveness.
\item \link{grades}: Fictitious data about statistics grades.
\item \link{ice_bucket}: Data about the ice bucket challenge.
\item \link{invisibility_base}: Fictitious data about how much mischief people would get up to if they had an invisibility cloak using a pre-post study design.
\item \link{invisibility_cloak}: Fictitious data about how much mischief people would get up to if they had an invisibility cloak using an independent design.
\item \link{invisibility_rm}: Fictitious data about how much mischief people would get up to if they had an invisibility cloak but using a repeated measures design.
\item \link{hangover}: fictitious data about the efficacy of different drinks as cures for a hangover.
\item \link{hiccups}: Fictitious data on digital rectal stimulation and hiccups.
\item \link{honesty_lab}: Fictitious data about perceptions of honesty.
\item \link{jiminy_cricket}: Fictitious data about whether wishing on a star makes you successful.
\item \link{johns_2012}: Data about whether the colour red is a mating signal to men.
\item \link{lambert_2012}: Data about whether pornography use is related to relationship commitment and infidelity.
\item \link{massar_2012}: Data about whether gossiping has an evolutionary function.
\item \link{mcnulty_2008}: Simulated data to match the results of a study about whether attractiveness is linked to the support given within a relationship.
\item \link{men_dogs}: Fictitious data about whether men exhibit dog-like behaviours (compared to dogs).
\item \link{metal_health}: Fictitious data about whether listening to heavy metal negatively affects mental health.
\item \link{metallica}: Data about thrash metal band Metallica.
\item \link{murder}: Fictitious data about the number of murder each month at three street locations (Ruskin Avenue, Acacia Avenue and Rue Morgue).
\item \link{muris_2008}: Data about whether you can train children to interpret ambiguous situations in a particular way.
\item \link{notebook}: Fictitious data about whether watching the film the notebook is emotionally arousing.
\item \link{ong_2011}: Data about social media profile pictures and personality traits.
\item \link{ong_tidy}: Data about social media profile pictures and personality traits.
\item \link{profile_pic}: Fictitious data related to whether the number of friend requests from random people on social media is affected by whether your profile picture depicts you as single or part of a romantic couple.
\item \link{pubs}: Data illustrating the difference between an outlier and an influential case.
\item \link{puppies}: Fictitious data related to whether puppy therapy works.
\item \link{puppy_love}: Fictitious data related to whether puppy therapy works when you adjust for a person's love of puppies.
\item \link{r_exam}: Fictitious data relating to an R exam at two universities.
\item \link{roaming_cats}: Fictitious data about how far cats roam from their homes.
\item \link{santas_log}: Fictitious data related to whether the type and quantity of treat consumed on Christmas night affects whether elves successfully deliver presents.
\item \link{self_help}: Fictitious data about whether self-help books improve relationship satisfaction.
\item \link{self_help_dsur}: Fictitious data about whether self-help books improve relationship satisfaction compared to statistics books.
\item \link{shopping}: Fictitious data about shopping.
\item \link{sniffer_dogs}: Fictitious data about training sniffer dogs to detect alien space lizards when they try to mask their identity with different scents. See also \link{alien_scents}.
\item \link{social_anxiety}: Fictitious (I think) data about whether social anxiety symptoms are specific to social anxiety.
\item \link{social_media}: Fictitious data about the effects of social media on grammar.
\item \link{soya}: fictitious data about the effects of eating soya on sperm count.
\item \link{speed_date}: Fictitious data related to the extent to which interest in dating someone is affected by their looks, personality or the dating strategy they adopt.
\item \link{stalker}: fictitious data about therapy for stalking.
\item \link{students}: I can't even remember what this data file contains.
\item \link{superhero}: fictitious data about whether wearing different superhero costumes leads to more severe physical injuries.
\item \link{supermodel}: Fictitious data about supermodel salaries.
\item \link{switch}: Fictitious data relating to whether injuries from playing video console games can be mitigated by a warm up.
\item \link{tablets}: Fictitious data about predicting the desirability of computing tablets.
\item \link{tea_15}: Fictitious data based on real data about cognitive functioning and drinking tea.
\item \link{tea_716}: Fictitious data based on real data about cognitive functioning and drinking tea.
\item \link{teaching}: Fictitious data about the success of different methods of teaching.
\item \link{teach_method}: More fictitious data about the success of different methods of teaching.
\item \link{tumour}: fictitious data about mobile phone use and brain tumours.
\item \link{video_games}: Fictitious data about the relationship between video game use, callous unemotional traits and aggression.
\item \link{xbox}: Fictitious data relating injuries to the type of video console game played and the console it was played on.
\item \link{tuk_2011}: Data about whether needing to urinate helps decision making.
\item \link{zhang_sample}: Data about whether performing a maths test under a different name assists performance.
}
}

\section{Smart Alex solutions}{


Solutions for end of chapter tasks are available at \href{https://www.discovr.rocks/solutions/alex/}{www.discovr.rocks/solutions/alex/}.
}

\section{Labcoat Leni solutions}{


Solutions for the Labcoat Leni tasks are available at \href{https://www.discovr.rocks/solutions/leni/}{www.discovr.rocks/solutions/leni/}.
}

\section{Chapter code}{


Although I recommend working through the interactive solutions, each book Chapter has online code and a downloadable R Markdown file available from \href{https://www.discovr.rocks/solutions/code/}{www.discovr.rocks/solutions/leni/}.
}

\section{References}{

\itemize{
\item Field, A. P. (2021). \emph{Discovering statistics using R and RStudio} (2nd ed.). London: Sage.
}
}

