% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{graphicGen}
\alias{graphicGen}
\title{splitts calculation in paired and unpaired sections}
\usage{
graphicGen(x, method, colNames, dataSetName)
}
\arguments{
\item{method}{is the method parameter and the}

\item{input}{is the data.frame}
}
\value{
returns the statistical calculations for each section
}
\description{
splitts calculation in paired and unpaired sections
}

