% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{johns_2012}
\alias{johns_2012}
\title{Johns et al. (2012) data}
\format{A tibble with 160 rows and 4 variables.}
\source{
\url{https://www.discoveringstatistics.com/books/discovering-statistics-using-r/}
}
\usage{
johns_2012
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
It is believed that males have a biological predispoition towards the colour red because it is sexually salient.
The theory suggests that women use the colour red as a proxy signal for genital colour to indicate ovulation and sexual proceptivity.
If this hypothesis is true then using the colour red in this way would have to attract men (otherwise it’s a pointless strategy).
In a novel study, Johns, Hargrave, and Newton-Fisher (2012) tested this idea by manipulating the colour of four pictures of female
geneitalia to make them increasing shades of red (pale pink, light pink, dark pink, red). Heterosexual males rated the resulting 16
pictures from 0 (unattractive) to 100 (attractive). These are the data from that study. The data contains the following variables:

\itemize{
\item \strong{id}: participant id
\item \strong{partners}: sexual experience coded as a factor ('Very little' and 'Some")
\item \strong{colour}: colour of the female geneitalia in image
\item \strong{attractiveness}: male rating of the attractiveness of the female geneitalia from 0 to 100
}
}
\keyword{datasets}
