% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{notebook}
\alias{notebook}
\title{The notebook data}
\format{
A tibble with 40 rows and 3 variables.
}
\source{
\url{https://www.discovr.rocks}
}
\usage{
notebook
}
\description{
A dataset from Field, A. P. (2020). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
Fictitious data about the film The Notebook. Imagine that a film company director was interested in whether there was really such a thing as a ‘chick flick’
(a film that has the stereotype of appealing to women more than to men). He took 20 men and 20 women and showed half of each sample a film that was supposed to be a ‘chick flick’ (The Notebook).
The other half watched a documentary about notebooks as a control. In all cases the company director measured participants’ arousal  as an indicator of how much they enjoyed the film.
The data contains the following variables:
\itemize{
\item \strong{sex}: biological sex of the participant
\item \strong{film}: whether the person watched The Notebook or a documentary about notebooks
\item \strong{arousal}: the person's average physiological arousal (e.g., emotional response) during the film.
}
}
\keyword{datasets}
