% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{pairedTest}
\alias{pairedTest}
\title{Output for only paired tests}
\usage{
pairedTest(input)
}
\arguments{
\item{input}{taking the output of all statistical methods}
}
\value{
extacts p value of all paired tests and adds them to data.frame
}
\description{
Output for only paired tests
}

