% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{presetFunction}
\alias{presetFunction}
\title{function for the preset usage}
\usage{
presetFunction(presetApp, x)
}
\arguments{
\item{presetApp}{is the input for the specific preset}
}
\value{
returns modified data output specific for the choosen preset
}
\description{
function for the preset usage
}

