% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{shapiroTest}
\alias{shapiroTest}
\title{Shapiro Wilks Test}
\usage{
shapiroTest(input)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
The calculation output
}
\description{
Shapiro Wilks Test
}

