% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr_data.R
\docType{data}
\name{teaching}
\alias{teaching}
\title{Method of teaching data (2 groups)}
\format{
A tibble with 20 rows and 3 variables.
}
\source{
\href{https://www.discovr.rocks/csv/teaching.csv}{www.discovr.rocks/csv/teaching.csv}
}
\usage{
teaching
}
\description{
A dataset from Field, A. P. (2026). Discovering statistics using R and RStudio (2nd ed.). London: Sage.
}
\details{
The data show the score (out of 20) for 20 different students, some of whom are biologically male and others biologically female,
and some of whom were taught using positive reinforcement (being nice) and others who were taught using punishment (electric shock)
\itemize{
\item \strong{id}: participant ID
\item \strong{method}: The type of teaching method used
\item \strong{sex}: Biological sex of the individual
\item \strong{mark}: The score out of 20 on a test
}
}
\keyword{datasets}
