% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{unPairedTest}
\alias{unPairedTest}
\title{calculating un-paired tests}
\usage{
unPairedTest(input)
}
\arguments{
\item{input}{is the data.frame}
}
\value{
returns the statistical calculations for each section
}
\description{
calculating un-paired tests
}

