% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticFun.R
\name{welchTest}
\alias{welchTest}
\title{Statistical Functions}
\usage{
welchTest(input)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
indicates whether parameter are from same population
}
\description{
Statistical Functions
}

