% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticFun.R
\name{welchTest}
\alias{welchTest}
\title{Statistical Functions}
\usage{
welchTest(mat, ...)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
indicates whether parameter are from same population - with different variances. 
the logic of the function syntax was fount at http://www.sthda.com/english/wiki/matrix-of-student-t-test
}
\description{
Statistical Functions
}

