% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovr.R
\name{wilcoxonTest}
\alias{wilcoxonTest}
\title{Wilcoxon Test}
\usage{
wilcoxonTest(input)
}
\arguments{
\item{input}{A data.frame or data.table}
}
\value{
Alternative test for the paired t-test if data is not normal - on dependend samples
}
\description{
Wilcoxon Test
}

